% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{add_individual_fl_data_benthic}
\alias{add_individual_fl_data_benthic}
\title{Add fork length data for individuals counted in maxn but with no available fork length
for these individuals we assign the mean fork length calculated in maxn data
for benthic bruvs}
\usage{
add_individual_fl_data_benthic(dat_fl, dat_maxn)
}
\arguments{
\item{dat_fl}{}

\item{dat_maxn}{}
}
\description{
Add fork length data for individuals counted in maxn but with no available fork length
for these individuals we assign the mean fork length calculated in maxn data
for benthic bruvs
}
