% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{add_mean_fl_to_maxn_data_benthic}
\alias{add_mean_fl_to_maxn_data_benthic}
\title{Add mean fork length data to maxn data based on a hierarchy for benthic bruvs}
\usage{
add_mean_fl_to_maxn_data_benthic(dat_fl, dat_maxn)
}
\arguments{
\item{dat_fl}{}

\item{dat_maxn}{}
}
\description{
Add mean fork length data to maxn data based on a hierarchy for benthic bruvs
}
