% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{add_mean_fl_to_maxn_data_pelagic}
\alias{add_mean_fl_to_maxn_data_pelagic}
\title{Add mean fork length data to maxn data based on a hierarchy for pelagic bruvs}
\usage{
add_mean_fl_to_maxn_data_pelagic(dat_fl, dat_maxn)
}
\arguments{
\item{dat_fl}{}

\item{dat_maxn}{}
}
\description{
Add mean fork length data to maxn data based on a hierarchy for pelagic bruvs
}
