% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{assign_empty_opcodes_meta_pelagic}
\alias{assign_empty_opcodes_meta_pelagic}
\title{Assign empty opcodes in meta for pelagic bruvs}
\usage{
assign_empty_opcodes_meta_pelagic(dat_meta, dat_maxn)
}
\arguments{
\item{dat_meta}{}

\item{dat_maxn}{}
}
\description{
Assign empty opcodes in meta for pelagic bruvs
}
