% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_response_variables.R
\name{beta_slope_benthic}
\alias{beta_slope_benthic}
\title{beta_slope_benthic}
\usage{
beta_slope_benthic(dat, dat2)
}
\arguments{
\item{dat}{}

\item{dat2}{}
}
\value{
data_to_export
}
\description{
beta_slope_benthic
}
\examples{

}
