% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_response_variables.R
\name{beta_slope_pelagic}
\alias{beta_slope_pelagic}
\title{Make beta slope}
\usage{
beta_slope_pelagic(dat, dat2)
}
\arguments{
\item{dat}{mean median and max responses}
}
\value{
dat3
}
\description{
Make beta slope
}
\examples{



}
