% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_response_variables.R
\name{beta_slope_pelagic_alt}
\alias{beta_slope_pelagic_alt}
\title{Make beta slope and alternative slope method}
\usage{
beta_slope_pelagic_alt(dat, dat2)
}
\arguments{
\item{dat}{mean median and max responses}
}
\value{
dat3
}
\description{
Make beta slope and alternative slope method
}
\examples{



}
