% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{bin_global_points_lm}
\alias{bin_global_points_lm}
\title{Title}
\usage{
bin_global_points_lm(dat2, minsize, lat_band)
}
\arguments{
\item{minsize}{}

\item{dat}{}
}
\value{
plot of bing_global and lm coefs
}
\description{
Title
}
\examples{


}
