% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{bin_global_points_lm_lon}
\alias{bin_global_points_lm_lon}
\title{Title}
\usage{
bin_global_points_lm_lon(dat, minsize, lon_band)
}
\arguments{
\item{dat}{}

\item{minsize}{}
}
\value{
plot of bing_global and lm coefs
}
\description{
Title
}
\examples{


}
