% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{bind_pelagic}
\alias{bind_pelagic}
\title{Bind pelagic (maxn or fl) data}
\usage{
bind_pelagic(dat1, dat2)
}
\arguments{
\item{dat1}{}

\item{dat2}{}
}
\value{

}
\description{
Bind pelagic (maxn or fl) data
}
