% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_weight.R
\name{bins_assignment_plot}
\alias{bins_assignment_plot}
\title{Demonstration of how binned body-mass values get assigned to logarithmic
size-class bins (Figure 3 of MEPS)}
\usage{
bins_assignment_plot(
  num2 = 11,
  lenBins = seq(10, 40, by = 5),
  LWa = c(0.001, 0.0255),
  LWb = c(3.4362, 2.7643),
  col1 = c("red", "pink"),
  col2 = c("blue", "lightblue"),
  thick = 7,
  xaxs = "i",
  yaxs = "i",
  mgp = c(2, 0.5, 0),
  lend = "butt"
)
}
\arguments{
\item{num2}{Number of log2 bin breaks}

\item{lenBins}{Example length bin breaks to use for the example body-mass bins}

\item{LWa}{Vector of two alpha values (see MEPS equation 1), for species 1
then species 2, though only species 2 is used, just keeping consistency
with \code{length_weight_plot()}}

\item{LWb}{Vector of two beta values (see MEPS equation 1), for species 1
then species 2, though only species 2 is used, just keeping consistency
with \code{length_weight_plot()}}

\item{col1}{Vector of two colours to alternate for species 1}

\item{col2}{Vector of two colours to alternate for species 2}

\item{thick}{Thickness of bin segments}

\item{xaxs, yaxs, mgp, lend}{standard options for \code{par()}, defaults are for
Figure 3 of MEPS}
}
\value{
Plots figure, default is Figure 3 of MEPS
}
\description{
Demonstration of how binned body-mass values get assigned to logarithmic
size-class bins (Figure 3 of MEPS)
}
\author{
Andrew Edwards
}
