% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{bruvs_protect_var}
\alias{bruvs_protect_var}
\title{plot predictor variables for each bruvs_protection_use cat}
\usage{
bruvs_protect_var(dat, envar_name, vert_value)
}
\arguments{
\item{dat}{data}

\item{envar_name}{envar variables}
}
\value{

}
\description{
plot predictor variables for each bruvs_protection_use cat
}
