% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{bruvs_var_range}
\alias{bruvs_var_range}
\title{plot predictor variables for each bruvs type}
\usage{
bruvs_var_range(dat, envar_name, axis_name)
}
\arguments{
\item{dat}{data}

\item{envar_name}{envar variables}
}
\value{

}
\description{
plot predictor variables for each bruvs type
}
