% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{calcLikeSegments}
\alias{calcLikeSegments}
\title{Calculate MLEs and 95\\% confidence intervals of \code{b} for separate body-mass segments
of a data set}
\usage{
calcLikeSegments(
  p = -1.5,
  w = NULL,
  d = NULL,
  bin_tibble = NULL,
  segmentIndices,
  ...
)
}
\arguments{
\item{p}{initial value of \code{b} for the numerical optimisation}

\item{w}{vector of length \code{J+1} giving the bin breaks \verb{w_1, w_2, ..., w_\{J+1\}}}

\item{d}{vector of length \code{J} giving the count in each bin; must have \code{d_1},
\code{d_J > 0}}

\item{bin_tibble}{data as a tibble, with each row representing a bin and
columns:
\itemize{
\item \code{wmin}: minimum body mass of that bin
\item \code{wmax}: maximum body mass of that bin
\item \code{binCount}: count in that bin
}}

\item{segmentIndices}{the indices of \code{w} or \code{bin_tibble$wmin} to use as breakpoints to separate the
range of body masses into distinct segments to be fit separately; must
have first element 1 and last element the value of \code{J+1}. MAYBE CHANGE?
CHECK ME: Segments are assigned based on minima of bins being \verb{>= w[segmentIndices]}, so we end up with \code{S = length(segmentIndices) - 1}
segments. Thus, \code{w[segmentIndices[i]]} is the minimum of segment \code{i}.}

\item{...}{further inputs to negLL.PLB.binned *** TODO maybe?}
}
\value{
list containing:
\itemize{
\item bins_in_segs: tibble with a row for each bin and columns (this is
\code{bin_tibble} with the extra column \code{segment}):
\itemize{
\item \code{wmin}: minimum body mass of that bin
\item \code{wmax}: maximum body mass of that bin
\item \code{binCount}: count in that bin
\item \code{segment}: which segment the bin falls in, an integer from 1 to S
}
\item b_segs: tibble of results with a row for each segment and columns:
\itemize{
\item \code{segment}: integer indicating the segment being fitted
\item \code{segMin}: mininum body mass of that segment
\item \code{segMax}: maximum body mass of that segment
\item \code{confMin}: minimum of 95\\% confidence interval of \code{b} for that segment
\item \code{b}: MLE for \code{b} for that segment
\item \code{confMax}:  maximum of 95\\% confidence interval of \code{b} for that segment
}
}
}
\description{
Rather than trying to estimate a single MLE of \code{b} across an entire data
set, it may make more sense fitting the PLB distribution independently
across segments (distinct ranges) of the data. For example, if the full
range of body sizes are not collected using the same sampling protocols,
such as when combining phytoplankton and zooplankton data. Data are input as
EITHER vectors \code{w} and \code{d} of bin breaks and bin counts, OR as a tibble with
one row for each bin and columns as described below.
}
\author{
Andrew Edwards
}
