% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{clean_data_with_vars}
\alias{clean_data_with_vars}
\title{clean data with predictor variables **ATTENTION need to update with new variables}
\usage{
clean_data_with_vars(dat, response_name)
}
\arguments{
\item{dat}{dataframe with predictor variables}
}
\value{

}
\description{
clean data with predictor variables **ATTENTION need to update with new variables
}
