% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{clean_maxn_benthic}
\alias{clean_maxn_benthic}
\title{Clean maxn data for benthic bruvs}
\usage{
clean_maxn_benthic(dat)
}
\arguments{
\item{dat}{}
}
\description{
Clean maxn data for benthic bruvs
}
