% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{clean_maxn_pelagic}
\alias{clean_maxn_pelagic}
\title{Clean maxn data for pelagic bruvs}
\usage{
clean_maxn_pelagic(dat)
}
\arguments{
\item{dat}{}
}
\description{
Clean maxn data for pelagic bruvs
}
