% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{clean_meta_benthic}
\alias{clean_meta_benthic}
\title{#' Read a and b parametersfrom fishbase
#'
#'
#' @return
#' @export}
\usage{
clean_meta_benthic(dat)
}
\arguments{
\item{dat}{}
}
\value{

}
\description{
read_a_b_parameters <- function(){

readxl::read_excel(here::here("data", "fishbase", "a_b_parameters.xlsx"),
sheet = 1)

}
Clean meta data for benthic bruvs
}
