% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{clean_meta_pelagic}
\alias{clean_meta_pelagic}
\title{Clean meta data for pelagic bruvs}
\usage{
clean_meta_pelagic(dat)
}
\arguments{
\item{dat}{}
}
\description{
Clean meta data for pelagic bruvs
}
