% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{coef_plot}
\alias{coef_plot}
\title{get coefficient plot for model}
\usage{
coef_plot(model, model_name)
}
\arguments{
\item{model}{gls}

\item{model_name}{}
}
\value{

}
\description{
get coefficient plot for model
}
