% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{coef_plot_signif_terms}
\alias{coef_plot_signif_terms}
\title{get coefficient plot for model for significant terms}
\usage{
coef_plot_signif_terms(model, model_name, signif_cutoff)
}
\arguments{
\item{model}{gls}

\item{model_name}{model name}

\item{signif_cutoff}{significance cutoff (eg 0.001)}
}
\value{

}
\description{
get coefficient plot for model for significant terms
}
