% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{coef_plot_terms}
\alias{coef_plot_terms}
\title{get coefficient plot for model for selection terms}
\usage{
coef_plot_terms(model, model_name)
}
\arguments{
\item{model}{gls}

\item{model_name}{}
}
\value{

}
\description{
get coefficient plot for model for selection terms
}
