% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{compare_gls_basic_cor_secondmode}
\alias{compare_gls_basic_cor_secondmode}
\title{compare gls models with secondmode as response variable and bruvs as factor of all covariates
with different basic autocorrelation structures (form=~1)  **ATTENTION need to update with new variables}
\usage{
compare_gls_basic_cor_secondmode(data)
}
\arguments{
\item{data}{dataframe with predictor variables}
}
\value{

}
\description{
compare gls models with secondmode as response variable and bruvs as factor of all covariates
with different basic autocorrelation structures (form=~1)  **ATTENTION need to update with new variables
}
