% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{compare_gls_spatial_cor_betaslope}
\alias{compare_gls_spatial_cor_betaslope}
\title{compare gls models with betaslope as response variable and bruvs as factor of all covariates
with different spatial autocorrelation structures (form=~long+lat)  **ATTENTION need to update with new variables}
\usage{
compare_gls_spatial_cor_betaslope(data)
}
\arguments{
\item{data}{dataframe with predictor variables}
}
\value{

}
\description{
compare gls models with betaslope as response variable and bruvs as factor of all covariates
with different spatial autocorrelation structures (form=~long+lat)  **ATTENTION need to update with new variables
}
