% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{conceptual_dia}
\alias{conceptual_dia}
\title{conceptual diagram with model and size spectra fishing effects}
\usage{
conceptual_dia(dat2, dat, data, min_size, bandw, scale, alpha)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
conceptual diagram with model and size spectra fishing effects
}
