% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{conceptual_marg_mod}
\alias{conceptual_marg_mod}
\title{combined marginal plots with conceptual diagram with model and size spectra fishing effects}
\usage{
conceptual_marg_mod(firstmode_marg, secondmode_marg, beta_marg, covariate_name)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
combined marginal plots with conceptual diagram with model and size spectra fishing effects
}
