% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{confPlot}
\alias{confPlot}
\title{Plot confidence intervals of repeated estimates for one method}
\usage{
confPlot(
  repConf,
  legName,
  b.true = b.known,
  inCol = "darkgrey",
  outCol = "blue",
  vertCol = "red",
  pchVal = 20,
  cexVal = 0,
  xLim = NULL,
  colourCode = TRUE,
  vertThick = 1,
  horizThick = 0.3,
  thin = 33,
  horizLines = FALSE,
  horizLinesOut = TRUE,
  horizLinesIn = TRUE,
  yLab = "Sample number",
  yTicks = seq(0, 300, 50),
  yLabels = TRUE,
  vertFirst = FALSE,
  insetVal = c(-0.08, -0.06),
  insetVal2 = c(-0.08, 0.07),
  xsmallticks = NULL,
  legLoc = "topleft"
)
}
\arguments{
\item{repConf}{data frame with columns \code{confMin} and \code{confMax}, the
minimum and maximum of the 95\\% confidence interval for b, and rows
corresponding to each simulated data set. When \code{confPlot} is called, for
some methods \code{b = slope - 1} or \code{b = slope - 2}, which gets done in the
call.}

\item{legName}{legend name for that panel}

\item{b.true}{true known value of b (as used for simulating the data)}

\item{inCol}{colour for intervals that \code{b.true} is inside}

\item{outCol}{colour for intervals that \code{b.true} is outside}

\item{vertCol}{colour of vertical line for \code{b.true}}

\item{pchVal}{\code{pch} for points at endpoints of intervals}

\item{cexVal}{size of points (default of 0 does not plot them)}

\item{xLim}{x-axis range}

\item{colourCode}{colour code the figure}

\item{vertThick}{thickness of vertical line for \code{b.true}}

\item{horizThick}{thickness of horizontal lines}

\item{thin}{number of values to thin the values for plotting. Only works for
33 or 99 for now (since they are factors of 9999, the value used in simulations).}

\item{horizLines}{whether to plot horizontal grey lines or not as example
(not needed now since doing lines for intervals)}

\item{horizLinesOut}{whether to plot horizontal lines for
intervals for which true value is outside the interval}

\item{horizLinesIn}{whether to plot horizontal lines for
intervals for which true value is inside the interval}

\item{yLab}{label for y-axis}

\item{yTicks}{where to have tickmarks on y-axis}

\item{yLabels}{whether or not to label tickmarks on y-axis}

\item{vertFirst}{whether or not to plot vertical line first in order to see
the horizontal lines better (for LCD plot)}

\item{insetVal}{inset shift for naming the panel}

\item{insetVal2}{inset shift for printing observed coverage percentage}

\item{xsmallticks}{where to put unlabelled small tick marks on x-axis}

\item{legLoc}{where to put the legend, as the first argument in \code{legend()}}
}
\value{
plots a panel and returns what is plotted as a sorted data frame
}
\description{
Plot confidence intervals of the repeated estimates of \code{b} for one
method. Gets called eight times to produce Figure 4 of MEE paper and Figure 5 of
MEPS paper. Plots
horizontal lines for the intervals, colour coded as to whether or not they
include the true value of b.
}
\author{
Andrew Edwards
}
