% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_response_variables.R
\name{correct_bathy_benthic}
\alias{correct_bathy_benthic}
\title{correct bathymetry data with measured depth data}
\usage{
correct_bathy_benthic(dat)
}
\arguments{
\item{dat}{dataframe of benthic bruvs data with environmental variables}
}
\value{

}
\description{
correct bathymetry data with measured depth data
}
