% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{cum_dist_plot}
\alias{cum_dist_plot}
\title{Cumulative distribution plotting - MLE}
\usage{
cum_dist_plot(dat, minsize, lat_band)
}
\arguments{
\item{dat}{size data}

\item{minsize}{minimum lat side}

\item{lat_band}{number of lat bands}
}
\value{

}
\description{
Cumulative distribution plotting - MLE
}
