% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{eight.methods.plot}
\alias{eight.methods.plot}
\title{Plotting fits of eight methods for a single data set (MEE Figure 2)}
\usage{
eight.methods.plot(eight.results = eight.results)
}
\arguments{
\item{eight.results}{Output list from \code{eightMethodsMEE()}}
}
\value{
Figure of eight panels, one for each method
}
\description{
Panel plot of all eight methods in MEE paper, showing fit of each one.
Function may need editing to make more general.
}
\author{
Andrew Edwards
}
