% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{eightMethods}
\alias{eightMethods}
\title{Compute size-spectra exponents for a dataset using all eight methods}
\usage{
eightMethods(oneYear = 1980, dataForYear = dplyr::filter(data, Year ==
  oneYear), figName = "eightMethods")
}
\arguments{
\item{oneYear}{the year of data to use, from that in the multiple years contained
in \code{data}. TODO: that should presumably be \code{dataForYear}}

\item{dataForYear}{TODO: I had this for \code{data}: local data frame that has a
unique row for every combination of \code{Year}, \code{SpecCode} and \code{LngtClass},
with columns:
\itemize{
\item \code{Number}: number of observed individuals of that species in that
length class in that year
\item \code{bodyMass}: body mass representative of such an individual, as calculated
previously by \code{LWa * LngtClass ^ LWb}.
}}

\item{figName}{figure name, will get appended by \code{-Year} for each year, to
create a \code{.png} file for each year.}
}
\value{
data frame with one row for each method, with columns:
\itemize{
\item \code{Year}
\item \code{Method}
\item \code{b}: estimate of b from that method
\item \code{confMin}: lower end of 95\% confidence interval of \code{b} for that method
\item \code{confMax}: upper end of 95\% confidence interval of \code{b} for that method.
}

Also saves a \code{.png} figure called \code{paste(figName, "-", oneYear, ".png")}, of the
fits for each of the eight methods.
}
\description{
TODO: Developed and is called in \code{nSea15analysis.Snw}, which was modified
from what was in fitting2.r. May be specific to the IBTS data set.
}
\author{
Andrew Edwards
}
