% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{estimate_weight_from_length}
\alias{estimate_weight_from_length}
\title{Estimate weight from observed length for a taxa using rfishbase}
\usage{
estimate_weight_from_length(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Estimate weight from observed length for a taxa using rfishbase
}
