% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{estimate_weight_from_length_regression}
\alias{estimate_weight_from_length_regression}
\title{Estimate weight from observed length for a taxa using rfishbase using a and b estimates from regression analysis}
\usage{
estimate_weight_from_length_regression(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Estimate weight from observed length for a taxa using rfishbase using a and b estimates from regression analysis
}
