% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_dplyr.R
\name{eval.string.dplyr}
\alias{eval.string.dplyr}
\title{Internal function used by s_filter, s_select etc.}
\usage{
eval.string.dplyr(.data, .fun.name, ...)
}
\arguments{
\item{.data}{input}

\item{.fun.name}{one of dplyr's functions}

\item{...}{other arguments to the the dplyr function}
}
\description{
Internal functions
}
\author{
Sebastian Kranz (https://gist.github.com/skranz/9681509)
}
