% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{ex_data_lat_cdp_lon}
\alias{ex_data_lat_cdp_lon}
\title{extended data multiplot}
\usage{
ex_data_lat_cdp_lon(lat, cdp)
}
\arguments{
\item{lat}{weights lat plot}

\item{cdp}{cumulative density plot}
}
\value{

}
\description{
extended data multiplot
}
