% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{f}
\alias{f}
\title{Format x to have the specified number of decimal places}
\usage{
f(x, dec.points = 0)
}
\arguments{
\item{x}{TODO}

\item{dec.points}{TODO}
}
\value{
x with the specified number of decimal places, and comma for thousands
}
\description{
TODO: change points to places presumably
TODO: give a more descriptive name
Format x to have supplied number of decimal places, have
thousands seperated by commas.
Taken from our Pacific Hake assessment:
https://github.com/cgrandin/hake-assessment/blob/master/doc/r/r-functions/utilities.r
}
\author{
Chris Grandin
}
