% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{figridges_lon}
\alias{figridges_lon}
\title{Title fig ridges sensitivity long}
\usage{
figridges_lon(dat, min_size, lon_band)
}
\arguments{
\item{min_size}{minimum size in kg}

\item{lon_band}{number of longitudinal bans}

\item{data}{size data}
}
\value{

}
\description{
try this to include mode https://rpkgs.datanovia.com/ggpubr/reference/stat_central_tendency.html
}
\examples{

}
