% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{figridges_overlap}
\alias{figridges_overlap}
\title{Title fig ridges overlap}
\usage{
figridges_overlap(dat, min_size, lat_band, bandw)
}
\arguments{
\item{min_size}{minimum size in kg}

\item{lat_band}{number of latitudinal bans}

\item{data}{size data}
}
\value{

}
\description{
try this to include mode https://rpkgs.datanovia.com/ggpubr/reference/stat_central_tendency.html
}
\examples{

}
