% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{figridges_overlap_exped}
\alias{figridges_overlap_exped}
\title{Title fig ridges overlap by exped}
\usage{
figridges_overlap_exped(dat, min_size, bandw, scale, alpha)
}
\arguments{
\item{min_size}{minimum size in kg}

\item{data}{size data}

\item{lat_band}{number of latitudinal bans}
}
\value{

}
\description{
try this to include mode https://rpkgs.datanovia.com/ggpubr/reference/stat_central_tendency.html
}
\examples{

}
