% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_gls_sat_cor_betaslope}
\alias{fit_gls_sat_cor_betaslope}
\title{fit fully staturated gls models with betaslope as response variable and bruvs as factor of all covariates
with autocorrelation  **ATTENTION need to update with new variables}
\usage{
fit_gls_sat_cor_betaslope(data)
}
\arguments{
\item{data}{dataframe with predictor variables}
}
\value{

}
\description{
fit fully staturated gls models with betaslope as response variable and bruvs as factor of all covariates
with autocorrelation  **ATTENTION need to update with new variables
}
