% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_gls_sim_cor_firstmode}
\alias{fit_gls_sim_cor_firstmode}
\title{fit fully simplified gls models with firstmode as response variable and bruvs as factor of all covariates
with autocorrelation  **ATTENTION need to update with new variables}
\usage{
fit_gls_sim_cor_firstmode(data)
}
\arguments{
\item{data}{dataframe with predictor variables}
}
\value{

}
\description{
fit fully simplified gls models with firstmode as response variable and bruvs as factor of all covariates
with autocorrelation  **ATTENTION need to update with new variables
}
