% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_gls_sim_cor_median_maxsize}
\alias{fit_gls_sim_cor_median_maxsize}
\title{fit simplified gls models with median_maxSize as response variable and bruvs as factor of all covariates
with autocorrelation}
\usage{
fit_gls_sim_cor_median_maxsize(data)
}
\arguments{
\item{data}{dataframe with predictor variables}
}
\value{

}
\description{
fit simplified gls models with median_maxSize as response variable and bruvs as factor of all covariates
with autocorrelation
}
