% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{fl_lengthweight}
\alias{fl_lengthweight}
\title{length weight relationship with marginal density plot}
\usage{
fl_lengthweight(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
length weight relationship with marginal density plot
}
\examples{

}
