% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{fl_species_ord_marg}
\alias{fl_species_ord_marg}
\title{species rank order weights vertical}
\usage{
fl_species_ord_marg(data, lower.line, mid.line, upper.line, minsize)
}
\arguments{
\item{data}{weights}

\item{lower.line}{lower dashed line y- intercept}

\item{mid.line}{middle line y- intercept}

\item{upper.line}{upper dashed line y-intercept}
}
\value{

}
\description{
species rank order weights vertical
}
\examples{

}
