% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{fl_species_rank_order_quan}
\alias{fl_species_rank_order_quan}
\title{species rank order according to weight quantiles (upper and lower)}
\usage{
fl_species_rank_order_quan(data, lower.quan, upper.quan, breaks)
}
\arguments{
\item{data}{}

\item{lower.quan}{}

\item{upper.quan}{}
}
\value{

}
\description{
species rank order according to weight quantiles (upper and lower)
}
