% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{fl_species_rank_order_vert}
\alias{fl_species_rank_order_vert}
\title{species rank order weights horizontal}
\usage{
fl_species_rank_order_vert(data, lower.line, upper.line)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
species rank order weights horizontal
}
\examples{

}
