% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullResults}
\alias{fullResults}
\title{Full results from using each fitting method on each year of IBTS data set.}
\format{Data frame with 240 rows (one row for each of 30 years for each of
the 8 methods) and corresponding columns for each year-method combination:
\itemize{
\item Year: Year of data
\item Method: Method used
\item b: Estimated size-spectrum exponent $b$
\item confMin, confMax: Minimum and maximum of 95\% confidence interval of
$b$
\item stdErr: Standard error of estimate of $b$
}}
\source{
Vignette \code{MEPS_IBTS_2}.
}
\usage{
fullResults
}
\description{
Used to make MEPS Figure 1.
}
\keyword{datasets}
