% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gap.barplot.cust}
\alias{gap.barplot.cust}
\title{Customising \code{plotrix::gap.barplot} for a histogram with a gap in y-axis}
\usage{
gap.barplot.cust(y, gap = c(9, 980), xaxlab, xtics, yaxlab,
  ytics = c(seq(0, 8, by = 4), seq(980, 988, by = 4)), midpoints,
  breakpoints, xlim, ylim = c(0, 17),
  xlab = expression(paste("Values, ", italic(x))),
  ylab = "Count in each bin", horiz = FALSE, col = NULL, N = n,
  ...)
}
\arguments{
\item{y}{vector of data values}

\item{gap}{range of values to be left out}

\item{xaxlab}{labels for the x axis ticks}

\item{xtics}{position of the x axis ticks}

\item{yaxlab}{labels for the y axis ticks}

\item{ytics}{position of the y axis ticks}

\item{midpoints}{TODO}

\item{breakpoints}{TODO}

\item{xlim}{optional x limits for the plot}

\item{ylim}{optional y limits for the plot}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}

\item{horiz}{whether to have vertical or horizontal bars}

\item{col}{color(s) in which to plot the values}

\item{N}{TODO: to do with ticks I think}

\item{...}{arguments passed to 'barplot'.}
}
\value{
Barplot with a gap in the y-axis
}
\description{
For Figure 1 of MEE paper, to make a histogram (barplot) with a gap in the y-axis.
Customising \code{gap.barplot()} from the package plotrix by Jim Lemon and others
Several default options here are customised for the particular plot (and to
change a few of the defaults in gap.barplot) so the code would
require some modifiying to use more generally.
}
\details{
This function modifies \code{plotrix::gap.barplot()}, between 2nd Sept 2014 and
finalised here in October 2019. \code{plotrix} was written by Jim Lemon and
others and is available from CRAN at https://cran.r-project.org/web/packages/plotrix/index.html.
}
\author{
Andrew Edwards
}
