% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_gls_diagnostics}
\alias{get_gls_diagnostics}
\title{get diagnostic plots for gls model}
\usage{
get_gls_diagnostics(dat, model, model_name)
}
\arguments{
\item{dat}{dataframe with predictor variables}

\item{model}{gls model}

\item{model_name}{name of model}
}
\value{

}
\description{
get diagnostic plots for gls model
}
