% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_partial_plot}
\alias{get_partial_plot}
\title{get partial plots for given predictor of gls model  **ATTENTION need to update with new variables}
\usage{
get_partial_plot(dat, model, model_name, vars)
}
\arguments{
\item{dat}{dataframe with predictor variables}

\item{model}{gls model}

\item{model_name}{charecter name of gls model}

\item{vars}{vector of predictor varibles}
}
\value{

}
\description{
get partial plots for given predictor of gls model  **ATTENTION need to update with new variables
}
