% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{globalmap_port}
\alias{globalmap_port}
\title{Global map of ports}
\usage{
globalmap_port(world, mar, dat)
}
\arguments{
\item{world}{the output of \code{ggplot2::map_data()}}

\item{mar}{an sp object: world marine boundaries}

\item{dat}{coordinates of...}
}
\value{
A ggplot2 object
}
\description{
Global map of ports
}
\examples{
## ...
}
