% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{globalmap_sum}
\alias{globalmap_sum}
\title{Global map of BRUVS sampling effort without eez SUMMARY}
\usage{
globalmap_sum(world, meta_pb)
}
\arguments{
\item{world}{the output of \code{ggplot2::map_data()}}

\item{meta_pb}{coordinates of...}
}
\value{
A ggplot2 object
}
\description{
Pelagic in blue and benthic in yellow
}
\examples{
## ...
}
