% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{globalmap_sum_eez}
\alias{globalmap_sum_eez}
\title{Global map of BRUVS sampling effort with eez SUMMARY}
\usage{
globalmap_sum_eez(world, mar, meta_pb)
}
\arguments{
\item{world}{the output of \code{ggplot2::map_data()}}

\item{mar}{an sp object: world marine boundaries}

\item{meta_pb}{coordinates of...}
}
\value{
A ggplot2 object
}
\description{
Pelagic in blue and benthic in yellow
}
\examples{
## ...
}
