% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{hist_betaslope}
\alias{hist_betaslope}
\title{hist plot betaslope}
\usage{
hist_betaslope(dat)
}
\arguments{
\item{dat}{}
}
\value{
figure of a response variable
}
\description{
hist plot betaslope
}
