% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{hist_spectra_nonstack}
\alias{hist_spectra_nonstack}
\title{stacked histogram with weight bins}
\usage{
hist_spectra_nonstack(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
stacked histogram with weight bins
}
