% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{histAxes}
\alias{histAxes}
\title{Histogram axes for figures as in Figure 3 of MEE paper}
\usage{
histAxes(yBigTickLab = c(0, 2000, 4000), yBigTickNoLab = seq(0, 6500,
  1000), ySmallTick = seq(0, 6500, 500), cexAxis = 0.9,
  xsmallticks = seq(-3.5, 0.5, 0.1), xbigticks = seq(-3.5, 0.5, 0.5),
  vertCol = "red", vertThick = 1, b.known = -2)
}
\arguments{
\item{yBigTickLab}{y-axis big ticks to label}

\item{yBigTickNoLab}{y-axis big ticks to not label}

\item{ySmallTick}{y-axis small ticks (unlabelled)}

\item{cexAxis}{font size for axis labels, defaults are for
\code{MEE_reproduce_2.Rmd}.}

\item{xsmallticks}{where to put unlabelled small tick marks on x-axis}

\item{xbigticks}{where to put big tick marks on x-axis}

\item{vertCol}{vertCol colour of vertical line for \code{b.known}}

\item{vertThick}{thickness of vertical line for \code{b.known}}

\item{b.known}{known value of $b$}
}
\value{
Adds axes to existing histogram
}
\description{
Do the histogram axes in vignette \code{MEE_reproduce_2.Rmd} and later, since almost all
panels will have same axes. Not very flexible, so may need \code{histAxes2()}
to plot up to 10,000 (though that needs working on, or adapt this one).
}
\author{
Andrew Edwards
}
