% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{histAxes2}
\alias{histAxes2}
\title{Do histogram axes for Figure TODO}
\usage{
histAxes2()
}
\value{
Adds axes to existing histogram TODO: check
}
\description{
Do the histogram axes for, in particular, TODO \code{fitting3rep-n10000.r},
since the larger n sample size means affects the resulting histograms
of estimated b. Not very flexible, just doing for the one figure.
Copying what is used for Llin method (so could go back and use this function
throughout earlier code).
}
\author{
Andrew Edwards
}
