% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Tests if each value in a vector is a whole number}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{vector}

\item{tol}{tolerance}
}
\value{
vector of TRUE/FALSE corresponding to each element of \code{x}
}
\description{
Test if each value in a vector is a whole number. Taken from ?is.integer and
documentation is added here.
}
\author{
Andrew Edwards
}
