% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ISD_bin_plot}
\alias{ISD_bin_plot}
\title{Recommended plots of individual size distribution and fit for binned data}
\usage{
ISD_bin_plot(data.year, b.MLE, b.confMin, b.confMax, year = NA,
  xlim = NA, xmin = NA, xmax = NA, yScaling = 0.75,
  MLE.round = 2, xLabel.small = c(5, 50, 500, 5000), yBig.inc = 1000,
  ySmall.inc = 250, ySmall.tcl = -0.2,
  xLab = expression(paste("Body mass (", italic(x), "), g")),
  yLab = expression(paste("Number of ", values >= italic(x))),
  inset.a = c(0, 0), inset.year = c(0, 0.04), seg.col = "green",
  rect.col = "grey", fit.col = "red", fit.lwd = 2, conf.lty = 2,
  par.mfrow = c(2, 1), par.mai = c(0.4, 0.5, 0.05, 0.3),
  par.cex = 0.7, mgp.vals = c(1.6, 0.5, 0))
}
\arguments{
\item{data.year}{tibble containing columns Year, wmin, wmax, Number,
countGTEwmin, lowCount, highCount}

\item{b.MLE}{maximum likelihood estimate of \emph{b} (ideally from the MLEbins method)}

\item{b.confMin}{lower 95\% confidence limits of \emph{b}}

\item{b.confMax}{upper 95\% confidence limits of \emph{b}}

\item{year}{year of data to go into legend (use NA if not applicable)}

\item{xlim}{(soft) limits of x-axis. If NA then automatically uses the minimum
\code{wmin} and maximum \code{wmax} for that data set (so good to set globally when
doing multiple years).}

\item{xmin, xmax:}{values of \code{xmin} and \code{xmax} to plot the PLB curve}

\item{yScaling}{Scaling of y-minimum of y-axis. Axis can't go to zero on
log-log plot, but goes to the proportion \code{yScaling} (which is less than 1)
of the minimum value of counts greater than the highest \code{wmin} value. Do
such that can see the right-most bin in all plots.}

\item{MLE.round}{number of decimal places to show MLE of b on the top plot}

\item{xLabel.small}{which small tickmarks to label on the x-axis}

\item{yBig.inc}{increment for labelled big tickmarks on the unlogged y-axis}

\item{ySmall.inc}{increment for small unlabelled tickmarks on the y-axis}

\item{ySmall.tcl}{length of small y-axis tick marks - only for (a) maybe}

\item{xLab}{label for x-axis}

\item{yLab}{label for y-axis}

\item{inset.a}{how far to inset (a) and (b)}

\item{inset.year}{how far to inset the year}

\item{seg.col}{colour for horizontal green lines showing range of body sizes
for each bin}

\item{rect.col}{colour to fill in the rectangles for each bin}

\item{fit.col}{colour to plot the MLE curve, and those for the confidence intervals}

\item{fit.lwd}{line weight for MLE curve}

\item{conf.lty}{line type for two curves for the MLE confidence intervals}

\item{par.mfrow}{vector giving the layout of the figures (number of rows,
number of columns)}

\item{par.mai}{margin size in inches}

\item{par.cex}{magnification of plotting text and symbols relative to default}

\item{mgp.vals}{margin line for axis title, axis labels and axis line}
}
\value{
two-panel figure of the recommended plot of binned data and the
fitted individual size distribution, like Figures 7 and S.5-S.34 of the
MEPS paper. See the vignette \code{MEPS_IBTS_recommend} for explicit example.
}
\description{
Plots the individual size distribution and the fit from the MLEbins method,
with linear and then logarithmic y-axes, in the recommended way that takes
into account the bin structure of the data, as in Figures 7 and S.5-S.34 of
the MEPS paper.
}
\author{
Andrew Edwards
}
