% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ISD_bin_plot_nonoverlapping}
\alias{ISD_bin_plot_nonoverlapping}
\title{Recommended plots of individual size distribution and fit for binned data with non-overlapping bins}
\usage{
ISD_bin_plot_nonoverlapping(
  binValsTibble = NULL,
  binBreaks = NULL,
  binCounts = NULL,
  b.MLE,
  b.confMin,
  b.confMax,
  ...
)
}
\arguments{
\item{binValsTibble}{tibble of binned data with each row representing a bin
and with columns \code{binMin} and \code{binMmax} (min and max break of each bin)
and \code{binCount} (count in that bin), as in the \code{binVals} component of the
output of \code{binData}. \code{wmin} and \code{wmax} can also be used instead of
\code{binMin} and \code{binMax}. Extra columns are ignored.}

\item{binBreaks}{vector of bin breaks}

\item{binCounts}{vector of bin counts}

\item{b.MLE}{maximum likelihood estimate of \emph{b} (ideally from the MLEbin method)}

\item{b.confMin}{lower 95\\% confidence limits of \emph{b}}

\item{b.confMax}{upper 95\\% confidence limits of \emph{b}}

\item{...}{further arguments to be passed to \code{ISD_bin_plot()}}
}
\description{
For data in binned form (i.e. bin breaks and counts), plots the individual
size distribution and the fit from the MLEbin method (alreadly calculated)
with linear and then logarithmic y-axes. This is a simpler version of
the recommended plots in Figures 7 and S.5-S.34 of the MEPS paper; simpler
because if the bins are not overlapping (i.e. just one set of bin breaks,
not differing by species like in Fig. 7) then we do not have the grey boxes,
just the horizontal green lines. See vignette \code{MLEbin_recommend.Rmd}.
}
\details{
Bin breaks and counts are input as EITHER a single tibble \code{binValsTibble}
with each row representing a bin, OR as vectors \code{binBreaks} and
\code{binCounts}.

This function calls \code{ISD_bin_plot()} which is the original one for Fig 7 of
MEPS paper.
}
\author{
Andrew Edwards
}
