% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_response_variables.R
\name{join_measured_depth_benthic}
\alias{join_measured_depth_benthic}
\title{Join measured depth to benthic bruvs data with environmental variables}
\usage{
join_measured_depth_benthic(dat)
}
\arguments{
\item{dat}{dataframe of benthic bruvs data with environmental variables}
}
\value{

}
\description{
Join measured depth to benthic bruvs data with environmental variables
}
