% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{keep_opcode_in_meta}
\alias{keep_opcode_in_meta}
\title{keep only opcodes that are available in meta in data (fl or maxn) for pelagic or benthic bruvs}
\usage{
keep_opcode_in_meta(meta_dat, dat)
}
\arguments{
\item{meta_dat}{}

\item{dat}{}
}
\description{
keep only opcodes that are available in meta in data (fl or maxn) for pelagic or benthic bruvs
}
