% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{keep_opcode_in_meta_pelagic}
\alias{keep_opcode_in_meta_pelagic}
\title{keep only opcodes that are available in meta in data (fl or maxn) for pelagic bruvs}
\usage{
keep_opcode_in_meta_pelagic(meta_dat, dat)
}
\arguments{
\item{meta_dat}{}

\item{dat}{}
}
\value{

}
\description{
keep only opcodes that are available in meta in data (fl or maxn) for pelagic bruvs
}
