% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{LBmizbinsFun}
\alias{LBmizbinsFun}
\title{Calculate the bin breaks for the LBmiz method}
\usage{
LBmizbinsFun(beta, xmin, xmax, k)
}
\arguments{
\item{beta}{to be calculated, \code{log10(beta)} is the constant binwidth on
log10 scale. \code{beta} is solution to\preformatted{0 = beta^(k-2) * (2 * beta-1) - xmax/xmin

  = 2 * beta^(k-1) - beta^(k-2) - xmax/xmin
}}

\item{xmin}{minimum of data (lower bound of lowest bin)}

\item{xmax}{maximum of data (upper bound of highest bin)}

\item{k}{number of bins}
}
\value{
value to be minimised by \code{nlm()}
}
\description{
Calculate the bin breaks for the LBmiz (log binning as done by the package
mizer), given \code{xmin} and \code{xmax} (min and max of data) and the number of
bins, \code{k}. To be minimised by \code{nlm()} to calculate \code{beta}.
}
\author{
Andrew Edwards
}
