% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{LBN_bin_plot}
\alias{LBN_bin_plot}
\title{Biomass size spectrum plot for binned data demonstrating uncertainties}
\usage{
LBN_bin_plot(
  binValsTibble = NULL,
  binBreaks = NULL,
  binCounts = NULL,
  b.MLE,
  b.confMin,
  b.confMax,
  plot.binned.fitted = TRUE,
  log.xy = "xy",
  xLim = NA,
  yLim = NA,
  rect.col = "grey",
  logLabels = FALSE,
  xLabel.small = c(2, 5, 20, 50, 200, 500),
  yLabel.small = c(5, 50, 500),
  xLab = expression(paste("Body mass ", italic(x), "(g)")),
  yLab = "Normalised biomass",
  x.PLB = NA,
  legend = TRUE,
  leg.pos = "topright",
  inset = c(0, 0),
  leg.text = "(a)",
  ...
)
}
\arguments{
\item{binValsTibble}{tibble of binned data with each row representing a bin
and with columns \code{binMin} and \code{binMmax} (min and max break of each bin)
and \code{binCount} (count in that bin), as in the \code{binVals} component of the
output of \code{binData}. \code{wmin} and \code{wmax} can also be used instead of
\code{binMin} and \code{binMax}. Extra columns are ignored.}

\item{binBreaks}{vector of bin breaks}

\item{binCounts}{vector of bin counts}

\item{b.MLE}{maximum likelihood estimate of \emph{b} (ideally from the MLEbin method)}

\item{b.confMin}{lower 95\\% confidence limits of \emph{b}}

\item{b.confMax}{upper 95\\% confidence limits of \emph{b}}

\item{plot.binned.fitted}{if TRUE then also plot the estimated normalised
biomass in each bin for the MLE of \emph{b} and it's confidence limits}

\item{log.xy}{Which axes to log, for \code{plot(..., log = log.xy)}. So "xy" for
log-log axes, "x" for only x-axis logged, "" for both axes unlogged.}

\item{x.PLB}{vector of values to use to plot the fitted PLB curve; if NA then
automatically calculated}

\item{legend}{if TRUE then add legend}

\item{leg.pos}{position of legend, from "bottomright"', '"bottom"',
'"bottomleft"', '"left"', '"topleft"', '"top"', '"topright"', '"right"'
and '"center"'.}

\item{inset}{inset distance vector for legend}

\item{leg.text}{text for legend}

\item{...}{further arguments to be passed to \code{plot()} and
\code{plot_binned_fitted()}}
}
\value{
TODO should return a tibble of results
}
\description{
Biomass size spectrum plot for binned data, showing the bin widths
explicitly and the normalised biomass in
each bin (with resulting uncertainties). So extending MEE Fig. 6 for already
binned data, using a new approach motivated by MEPS Fig. 7. Should probably
then be recommended to replace MEE Fig. 6. Axes are logged, and labelled as either
logged or (more intuitive) logged values. See vignette \code{MLEbin_recommend.Rmd}.
}
\details{
Bin breaks and counts are input as EITHER a single tibble \code{binValsTibble}
with each row representing a bin, OR as vectors \code{binBreaks} and
\code{binCounts}.
}
\author{
Andrew Edwards
}
