% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{LBNbiom.method}
\alias{LBNbiom.method}
\title{Fit biomass size spectrum with the LBNbiom and LBbiom methods}
\usage{
LBNbiom.method(bodyMass = NULL, counts = NULL, binBreaks = NULL,
  lowerCutOff = 0)
}
\arguments{
\item{bodyMass}{vector of individual body masses}

\item{counts}{dataframe (or array) with first column being a body mass
value, and second column being the counts of the number of individuals for
that body mass. Only bodyMass or counts can be specified.}

\item{binBreaks}{breaks for the bins to be used to bin the data and then fit
the regression. If not provided then it calculates them as bin widths that
double in size (with breaks that are powers of 2) that encompass the data,
resulting in binBreaks   ..., 0.25, 0.5, 1, 2, 4, 8, 16,.... as
necessary.}

\item{lowerCutOff}{body mass representing the lower cut off for the range being
fit}
}
\value{
list containing:
\itemize{
\item \code{indiv}: dataframe with a row for each \code{bodyMass} value and with columns:
\itemize{
\item 'x': original \code{bodyMass} values.
\item \code{binMid}, \code{binMin}, \code{binMax}, \code{binWidth}: midpoint, minimum, maximum,
and width, respectively, of the bin within which the \code{x} value falls.
If \code{indiv} has >=10^6 rows then it is not saved.
If \code{counts} was specified then, for now, an equivalent \code{bodyMass}
vector was created and is column \code{x} (i.e. body masses are repeated).
}
\item binVals: dataframe with a row for each bin and columns:
\itemize{
\item \code{binMid}, \code{binMin}, \code{binMax}, \code{binWidth}: midpoint, minimum, maximum, and
width, respectively, of the bin.
\item \code{totalBiom}: total biomass in that bin
\item \code{totalBiomNorm}: normalised total biomass in that bin, defined as \code{totalBiom / binWidth}
\item \code{log10....}: \code{log10} of some of the above quantities
}
\item norm.lm: \code{lm()} result of the linear regression fit using normalised
biomass in each bin
\item norm.slope: slope of the linear regression fit using normalised
biomass in each bin
\item unNorm.lm: \code{lm()} result of the linear regression fit when not
normalising the biomass in each bin
\item unNorm.slope: slope of the linear regression fit when not
normalising the biomass in each bin
}
}
\description{
Use the log-binning with normalisation technique (LBNbiom method) to
calculate the slope of the biomass size spectra. Slope is from fitting
a linear regression of log10(normalised biomass in bin) against
log10(midpoint of bin). Bins can be defined by user, else are created to
double in size. Also calculates slope for the LBbiom method, for which
biomasses are not normalised.
}
\author{
Andrew Edwards
}
