% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{legJust}
\alias{legJust}
\title{Add legend with right-justification}
\usage{
legJust(textVec, pos = "topright", textWidth = "slope=-*.**",
  inset = 0, logxy = FALSE)
}
\arguments{
\item{textVec}{text for the legend, one element for each row}

\item{pos}{position of the legend (not tested for all positions)}

\item{textWidth}{width to make the text}

\item{inset}{inset distance}

\item{logxy}{TRUE if axes are logarithmic}
}
\value{
Adds legend to exiting plot. Returns NULL.
}
\description{
Add legend with right-justification, functionalising Uwe Ligges'
example in \code{?legend}. Really a way of adding text automatically in
the corner (which \code{legend()} works out the positioning for).
}
\examples{
\dontrun{ TODO check
  plot(10:1)
 legJust(c("Method", paste("value=", mean(1:10), sep=""), "x=7.0"))
}
}
\author{
Andrew Edwards
}
