% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_weight.R
\name{length_weight_plot}
\alias{length_weight_plot}
\title{Example length-weight relationships for two species, demonstrating
consequences of length bins}
\usage{
length_weight_plot(
  sp1 = "Common Ling",
  sp2 = "Lemon Sole",
  LWa = c(0.001, 0.0255),
  LWb = c(3.4362, 2.7643),
  col1 = c("red", "pink"),
  col2 = c("blue", "lightblue"),
  thick = 7,
  curve_thick = 3,
  length = 10:50,
  lenBins = seq(10, 40, by = 5),
  xlim = c(-6, 50),
  ylim = c(-60, 800),
  inset = c(0.1, -0.02),
  xlab = "Length, cm",
  ylab = "Body mass, g",
  xaxs = "i",
  yaxs = "i",
  mgp = c(2, 0.5, 0),
  lend = "butt"
)
}
\arguments{
\item{sp1}{Species 1 name (for legend)}

\item{sp2}{Species 2 name (for legend)}

\item{LWa}{Vector of two alpha values (see MEPS equation 1), for species 1
then species 2}

\item{LWb}{Vector of two beta values (see MEPS equation 1), for species 1
then species 2}

\item{col1}{Vector of two colours to alternate for species 1}

\item{col2}{Vector of two colours to alternate for species 2}

\item{thick}{Thickness of bin segments}

\item{curve_thick}{Thickness for curves}

\item{length}{Sequence of length values to plot the length-weight
relationships}

\item{lenBins}{Example length bin breaks to show as examples}

\item{xlim, ylim}{Limits for x and y axes}

\item{inset}{Inset for the legend}

\item{xlab, ylab}{Labes for x and y axes}

\item{xaxs, yaxs, mgp, lend}{standard options for \code{par()},
defaults are for Figure 2 of MEPS}
}
\value{
plots Figure, default is Figure 2 of MEPS
}
\description{
Reproduces Figure 2 of MEPS paper (for default values), showing how the same
length bins translate to different body-mass bins for different species. Has
not been fully tested with other input values (and some axis values etc are
still hardwired).
}
\author{
Andrew Edwards
}
