% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lengthToMass}
\alias{lengthToMass}
\title{Convert a vector of body lengths to a vector of body masses using length-weight
relationship}
\usage{
lengthToMass(lengths, LWa, LWb)
}
\arguments{
\item{lengths}{vector of lengths}

\item{LWa}{multiplicative coefficient in equation}

\item{LWb}{exponent in equation}
}
\value{
vector of body masses
}
\description{
Convert a vector of body lengths to a vector of body masses using the formula
\code{mass = LWa * lengths^LWb}. \code{LWa} and \code{LWb} will be species-specific.
User must ensure that units are consistent.
}
\author{
Andrew Edwards
}
