% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{Llin.method.counts}
\alias{Llin.method.counts}
\title{The Llin method for count data
The Llin method, which is plotting binned counts on log-linear
axes and then fitting a regression, for count data.}
\usage{
Llin.method.counts(valCounts, num.bins = NULL, binBreaks = NULL)
}
\arguments{
\item{valCounts}{valCounts: data.frame (can be tbl_df) with columns \code{bodyMass}
and \code{Number} (which is the count for each body mass). \code{bodyMass} can
represent midpoints, say, of existing bins, or be the actual
species-specific converted-to-bodyMass values. \code{Number} can be
non-integer, which can arise from standardising, say, trawl data to be per
hour.}

\item{num.bins}{number of bins to be used, though this is only a suggestion
since can get over-ridden by \code{hist()}. Need to specify \code{num.bins} OR \code{binBreaks}.}

\item{binBreaks}{breaks for the bins to be used to bin the data and
then fit the regression.}
}
\value{
list containing:
\itemize{
\item \code{mids}: midpoint of bins
\item \code{log.counts}: \code{log(counts)} in each bin
\item \code{counts}: counts in each bin
\item \code{lm}: results of the linear regression of \code{log.counts ~ mids}
\item \code{slope}: slope of the linear regression fit
\item \code{breaks}: bin breaks
\item \code{confVals}: 95\% confidence interval of the fitted slope
\item \code{stdError}: standard error of the fitted slope
}
}
\description{
The Llin method for count data
The Llin method, which is plotting binned counts on log-linear
axes and then fitting a regression, for count data.
}
\author{
Andrew Edwards
}
