% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{Llin.method}
\alias{Llin.method}
\title{Fit size spectrum using Llin method}
\usage{
Llin.method(bodyMass, num.bins = NULL, binBreaks = NULL)
}
\arguments{
\item{bodyMass}{vector of individual body masses}

\item{num.bins}{number of bins to be used, though this is only a suggestion
since can get over-ridden by \code{hist()}}

\item{binBreaks}{breaks for the bins to be used to bin the data and
then fit the regression}
}
\value{
list containing:
\itemize{
\item mids: midpoint of bins
\item log.counts: log(counts) in each bin
\item counts: counts in each bin
\item lm: results of the linear regression of \code{log.counts ~ mids}
\item slope: slope of the linear regression fit
\item breaks: bin breaks
\item confVals: 95\\% confidence interval of the fitted slope
}
}
\description{
Fit size spectrum using Llin method, which is plotting binned counts on
log-linear axes and then fitting a linear regression.
}
\author{
Andrew Edwards
}
