% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{lm.line}
\alias{lm.line}
\title{Plot bounded straight line of results of \code{lm()} fit}
\usage{
lm.line(x.vector, lm.results, ...)
}
\arguments{
\item{x.vector}{values of x (that have been fitted to), fitted line will be
bound by the range of \code{x.vector}}

\item{lm.results}{output from an \code{lm()} fit, with intercept \code{lm.results$coeff[1]}}

\item{...}{arguments to lines}
}
\value{
Adds a line to an existing plot
}
\description{
Plot a straight line between lowest and highest values of \code{x.vector}
based on \code{lm()} results, to use instead of \code{abline(lm.output)}
which extends beyond the fitted data.
}
\author{
Andrew Edwards
}
