% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{log2bins}
\alias{log2bins}
\title{Construct bins that double in size and encompass the data,}
\usage{
log2bins(x = NULL, counts = NULL)
}
\arguments{
\item{x}{vector of individual values (e.g. body masses)}

\item{counts}{dataframe (or array) with first column being an \code{x} value
(e.g. body mass), and second column being the \code{counts} of the number of
individuals for that value. Only \code{x} or \code{counts} can be specified.}
}
\value{
\preformatted{  list containing:
}
\itemize{
\item indiv: dataframe with a row for each x value, with columns:
\itemize{
\item \code{x}: original \code{x} values
\item \code{binMid}, \code{binMin}, \code{binMax}, \code{binWidth}: midpoint, minimum, maximum,
and width, respectively, of the bin within which the \code{x} value falls.   If
\code{indiv} has >=10^6 rows then it isn't saved.  If \code{counts} was specified
then an equivalent \code{x} vector is created and is column \code{x} (i.e. \code{x}
values are repeated). May not be the most efficient way, but it easiest to
program. May not need \code{indiv} returned, but it needs to be calculated
anyway.
}
\item \code{binVals}: dataframe with a row for each new bin, with columns:
\itemize{
\item \code{binMid}, \code{binMin}, \code{binMax}, \code{binWidth}: midpoint, minimum, maximum, and width,
respectively, of the bin
\item \code{binCount}: total number of individuals in that bin
\item \code{binCountNorm}: \code{binCount / binWidth}
\item \code{binSum}: sum of individual values in that bin (appropriate if \code{x}
represents biomass, but not length)
\item \code{binSumNorm}: \code{binSum / binWidth}
\item \code{log10...}: \code{log10()} of some of the above quantities.
}
}
}
\description{
Construct bins that double in size and encompass the data, resulting in
\code{binBreaks} ..., 0.25, 0.5, 1, 2, 4, 8, 16,.... as necessary, where the
breaks are powers of 2. Adapting from \code{LBNbiom.method()}. See \code{binData()}
for a generalised version.
}
\author{
Andrew Edwards
}
