% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{logTicks}
\alias{logTicks}
\title{Add axes and tick marks to a log-log plot to represent unlogged values}
\usage{
logTicks(xLim, yLim = NULL, tclSmall = -0.2, xLabelSmall = NULL,
  yLabelSmall = NULL, xLabelBig = NULL, mgpVal = c(1.6, 0.5, 0))
}
\arguments{
\item{xLim}{the x limits for the plot (unlogged scale); if NULL then do not
add anything to x-axis}

\item{yLim}{the y limits for the plot (unlogged scale); if NULL then
do not add anything to y-axis}

\item{tclSmall}{size of small tick marks}

\item{xLabelSmall}{which small tick marks on x-axis to label}

\item{yLabelSmall}{which small tick marks on y-axis to label}

\item{xLabelBig}{which big tick marks on the x-axis to label
(when automated they can overlap, so may need to specify)}

\item{mgpVal}{\code{mgp} values for axes. See \code{?par}}
}
\value{
Adds axes and big and small tick marks to the plot. Returns NULL
}
\description{
Useful because you can then interpret the unlogged values, e.g. Figures 2(h)
and 6(b) of MEE paper and Figure 7 of MEPS paper.
}
\examples{
\dontrun{
# Adapt the following (could make an explicit example):
  plot(..., log="xy", xlab=..., ylab=..., xlim=..., ylim=..., axes=FALSE)
  xLim = 10^par("usr")[1:2]
  yLim = 10^par("usr")[3:4]
  logTicks(xLim, yLim, xLabelSmall = c(5, 50, 500))
}
}
\author{
Andrew Edwards
}
