% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{make_correlogram_vars}
\alias{make_correlogram_vars}
\title{make correlogram of predictors ** ATTENTION need to update variables}
\usage{
make_correlogram_vars(dat)
}
\arguments{
\item{dat}{dataframe with predictor variables}
}
\value{

}
\description{
make correlogram of predictors ** ATTENTION need to update variables
}
