% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{make_map_meta}
\alias{make_map_meta}
\title{Make map of meta for pelagic/benthic bruvs}
\usage{
make_map_meta(meta, type)
}
\arguments{
\item{meta}{}

\item{type}{}
}
\description{
Make map of meta for pelagic/benthic bruvs
}
