% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{marg_plot_bathy}
\alias{marg_plot_bathy}
\title{marginal plots for of logged bathy variable}
\usage{
marg_plot_bathy(response, mod_name, dat, mod, var, var_name, group)
}
\arguments{
\item{dat}{dataframe with covariates}

\item{mod}{model of interest dataframe with predictor variables}

\item{var_name}{name of covariate of interest}
}
\value{

}
\description{
marginal plots for of logged bathy variable
}
