% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{marg_plot_bruvs_prot_rug}
\alias{marg_plot_bruvs_prot_rug}
\title{marginal plots for categorical co_variates with noextra but with bruvs as facet wrap and rug}
\usage{
marg_plot_bruvs_prot_rug(pred, var_name, rug_beta)
}
\arguments{
\item{pred}{data}

\item{var_name}{variable name}

\item{rug_beta}{rug name}
}
\value{

}
\description{
marginal plots for categorical co_variates with noextra but with bruvs as facet wrap and rug
}
