% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{marg_plot_bruvs_prot_sens}
\alias{marg_plot_bruvs_prot_sens}
\title{marginal sensitivity plots for categorical co_variates with noextra but with bruvs as facet wrap}
\usage{
marg_plot_bruvs_prot_sens(pred, var_name)
}
\arguments{
\item{pred}{data}

\item{var_name}{variable name}
}
\value{

}
\description{
marginal sensitivity plots for categorical co_variates with noextra but with bruvs as facet wrap
}
