% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{marg_plot_cat_covar}
\alias{marg_plot_cat_covar}
\title{marginal plots for categorical co_variates}
\usage{
marg_plot_cat_covar(
  response,
  mod_name,
  dat,
  mod,
  var,
  var_name,
  group,
  group2,
  condition
)
}
\arguments{
\item{dat}{dataframe with covariates}

\item{mod}{model of interest dataframe with predictor variables}

\item{var_name}{name of covariate of interest}
}
\value{

}
\description{
marginal plots for categorical co_variates
}
