% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{marg_plot_cat_covar_separate_bruvs}
\alias{marg_plot_cat_covar_separate_bruvs}
\title{marginal plots for categorical co_variates with noextra but with bruvs as facet wrap}
\usage{
marg_plot_cat_covar_separate_bruvs(
  response,
  mod_name,
  dat,
  mod,
  var,
  var_name,
  group,
  group2,
  condition
)
}
\arguments{
\item{response, }{dat with predictor variables}

\item{mod_name}{name of model}

\item{dat}{data}

\item{mod}{model}

\item{var}{variable}

\item{var_name}{variable name}

\item{group}{group}

\item{group2}{second group}
}
\value{

}
\description{
marginal plots for categorical co_variates with noextra but with bruvs as facet wrap
}
