% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{marg_plot_log}
\alias{marg_plot_log}
\title{marginal plots for the empirical covariate range of logged variable}
\usage{
marg_plot_log(response, mod_name, dat, mod, var, var_name, group, condition)
}
\arguments{
\item{dat}{dataframe with covariates}

\item{mod}{model of interest dataframe with predictor variables}

\item{var_name}{name of covariate of interest}
}
\value{

}
\description{
marginal plots for the empirical covariate range of logged variable
}
