% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{massToLength}
\alias{massToLength}
\title{Convert a vector of body masses to a vector of body lengths using length-weight
relationship (for simulations)}
\usage{
massToLength(masses, LWa, LWb)
}
\arguments{
\item{masses}{vector of masses}

\item{LWa}{multiplicative coefficient in equation}

\item{LWb}{exponent in equation}
}
\value{
vector of body lengths
}
\description{
Convert a vector of body masses to a vector of body lengths using the
formula
\code{length = (mass / LWa)^(1/b)}, which is the inverse of the standard
\code{mass = LWa * lengths^LWb}. \code{LWa} and \code{LWb} will be species-specific.
User must ensure that units are consistent.
}
\author{
Andrew Edwards
}
