% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_response_variables.R
\name{mean_median_max_benthic}
\alias{mean_median_max_benthic}
\title{Title}
\usage{
mean_median_max_benthic(dat)
}
\arguments{
\item{dat}{benthic sizes}
}
\value{

}
\description{
Title
}
\examples{



}
