% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_response_variables.R
\name{mean_median_max_pelagic}
\alias{mean_median_max_pelagic}
\title{Make mean, median and maximum pelagic response}
\usage{
mean_median_max_pelagic(dat)
}
\arguments{
\item{dat}{body size}
}
\value{
mean median and maximum values for pelagic sizes
}
\description{
Make mean, median and maximum pelagic response
}
\examples{



}
