% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{merge_fl_benthic_meta}
\alias{merge_fl_benthic_meta}
\title{merge benthic meta with benthic FL}
\usage{
merge_fl_benthic_meta(dat_meta, dat_fl)
}
\arguments{
\item{dat_meta}{...}

\item{dat_fl}{...}
}
\description{
merge benthic meta with benthic FL
}
