% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{merge_fl_pelagic_meta}
\alias{merge_fl_pelagic_meta}
\title{merge pelagic meta with pelagic FL}
\usage{
merge_fl_pelagic_meta(dat_meta, dat_fl)
}
\arguments{
\item{dat_meta}{...}

\item{dat_fl}{...}
}
\description{
merge pelagic meta with pelagic FL
}
