% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{MLE.plot}
\alias{MLE.plot}
\title{Recommended single plot of ISD and MLE results}
\usage{
MLE.plot(x, b, confVals = NULL, panel = FALSE, log.xy = "xy",
  mgpVals = c(1.6, 0.5, 0), inset = c(0, -0.04))
}
\arguments{
\item{x}{Vector of data values}

\item{b}{Estimate of b}

\item{confVals}{Confidence interval for estimate of b (two-component vector
for bounds of confidence interval)}

\item{panel}{Which panel number for multi-panel plots. \code{"h"} gives the
method and MLE estimate (as in MEE Figure 2(h)), \code{"b"} gives just (b) as
in Figure 6(b) and plots confidence interval curves, and NULL gives nothing.}

\item{log.xy}{Which axes to log, for \code{plot(..., log = log.xy)}. So "xy" for
log-log axes, "x" for only x-axis logged.}

\item{mgpVals}{mgp values to use, as in \code{plot(..., mgp = mgpVals)}.}

\item{inset}{Inset distance for legend}
}
\value{
Single figure of ISD on log-log plot (or log-linear depending on the
options given).
}
\description{
Gives Figure 2h and 6b of MEE.
}
\author{
Andrew Edwards
}
