% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{MLE.plots.recommend}
\alias{MLE.plots.recommend}
\title{Recommended LBN-type plot with ISD for fitted MLE (MEE Figure 6)}
\usage{
MLE.plots.recommend(x, b.MLE, confVals.MLE, hLBN.biom.list = NULL,
  inset = c(0, -0.04), mgpVals = c(1.6, 0.5, 0), xLim = c(0, 2.7),
  yLim = c(0, 3), yBigTicks = 0:3, ySmallTicks = c(0.5, 1.5, 2.5))
}
\arguments{
\item{x}{Raw data, must be body-mass values (else LBN plot is meaningless).}

\item{b.MLE}{MLE estimate of $b$}

\item{confVals.MLE}{Confidence interval for estimate of b (two-component vector
for bounds of confidence interval)}

\item{inset}{Inset distance for legend}

\item{LBNbiom.list}{Results from LBNbiom.method(x). If NULL then calculate
them here.}
}
\value{
Panel plot with LBN plot at top and ISD with fitted MLE and
confidence intervals at the bottom.
}
\description{
Only valid when \code{x} is biomass, since LBN plot calculates the biomass.
}
\author{
Andrew Edwards
}
