% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MLEbin.MEPS.cutoff16}
\alias{MLEbin.MEPS.cutoff16}
\title{Simulation results of multiple data sets (with $x_{min}=1$ but data sampled
above 16) binned using four binning
types and fit using MLEmid and MLEbin methods.}
\format{List containing two arrays and a list of parameters. See
\code{?sizeSpectra::MLEbin.simulate} for details.}
\source{
From running \code{MLEbin.simulate()} in \code{data-raw/simulate-data2.R}.
}
\usage{
MLEbin.MEPS.cutoff16
}
\description{
Similar to \code{MLEbin.MEPS.default} but with only sampling data above the cut
off value of 16. Produces results for
Figures S.37 and S.38 and Table S.5 of MEPS paper, as shown in
'MEPS_reproduce_2.Rmd' vignette.
}
\keyword{datasets}
