% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MLEbin.MEPS.default}
\alias{MLEbin.MEPS.default}
\title{Default simulation results of multiple data sets binned using four binning
types and fit using MLEmid and MLEbin methods.}
\format{
List containing two arrays and a list of parameters. See
\code{?sizeSpectra::MLEbin.simulate} for details.
}
\source{
From running \code{MLEbin.simulate()} in \code{data-raw/simulate-data2.R}.
}
\usage{
MLEbin.MEPS.default
}
\description{
Simulation results from generating 10,000 data sets from the PLB distribution
with default parameter values,
binning each data set using linear bins of width 1, 5 and 10, and also bins that
progressively double in width, and then fitting each data set using the
MLEmid and MLEbin likelihood methods.
As in Figures 4 and 5 and Table S.3 of MEPS paper. See 'MEPS_reproduce_2.Rmd'
vignette for code for those figures and tables, and  All simulated data sets have
the same parameters for PLB and the same sample size 'n'.
Individual data sets are not saved as they quickly take up a lot
of memory (would be 'num.reps' \eqn{\times} 'n' random numbers, which
for the default values is 10^7).
}
\keyword{datasets}
