% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MLEbin.MEPS.xmin16}
\alias{MLEbin.MEPS.xmin16}
\title{Simulation results of multiple data sets (with $x_{min}=16$) binned using four binning
types and fit using MLEmid and MLEbin methods.}
\format{List containing two arrays and a list of parameters. See
\code{?sizeSpectra::MLEbin.simulate} for details.}
\source{
From running \code{MLEbin.simulate()} in \code{data-raw/simulate-data2.R}.
}
\usage{
MLEbin.MEPS.xmin16
}
\description{
Similar to \code{MLEbin.MEPS.default} but with $x_{min}=16$. Produces results for
Figures S.35 and S.36 and Table S.4 of MEPS paper, as shown in
'MEPS_reproduce_2.Rmd' vignette.
}
\keyword{datasets}
