% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MLEbins.res}
\alias{MLEbins.res}
\title{Results from using MLEbins method on each year of IBTS data set, as in MEPS Table S.2}
\format{Table data frame (\code{tbl_df}) with 30 rows (one for each year) and
corresponding columns for each year:
\describe{
\item{Year}{Year of data}
\item{xmin}{Estimated $x_{min}$}
\item{xmax}{Estimated $x_{max}$}
\item{n}{Total number of individuals per hour}
\item{b}{Estimated size-spectrum exponent $b$}
\item{confMin, confMax}{Minimum and maximum of 95\% confidence interval of
$b$}
\item{C}{Calculated normalisation constant}
}}
\source{
Vignette \code{MEPS_IBTS_MLEbins}, TODO tidy up IBTS-MLEbins.R
}
\usage{
MLEbins.res
}
\description{
Results from using MLEbins method on each year of IBTS data set, as in MEPS Table S.2
}
\keyword{datasets}
