% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{MLEmid.MLEbin.conf}
\alias{MLEmid.MLEbin.conf}
\title{One figure with eight confidence interval plots for each of  MLEmid and MLEbin methods and four binning types}
\usage{
MLEmid.MLEbin.conf(results.list, vertCol = "red", vertThick = 1,
  xrange = c(-2.4, -1.1), xbigticks.by = 0.4, xsmallticks.by = 0.1,
  omi = c(0.2, 0.05, 0.22, 0.1), mfrow = c(4, 2), mai = c(0.3, 0.5,
  0.08, 0), xaxs = "i", yaxs = "i", mgp = c(2, 0.5, 0), cex = 0.8,
  inset = c(0, -0.04), insetMat = matrix(rep(c(-0.01, -0.04), 4), ncol
  = 2, byrow = TRUE), legLabMid = c("(a)", "(c)", "(e)", "(g)"),
  legLabBin = c("(b)", "(d)", "(f)", "(h)"))
}
\arguments{
\item{results.list}{output list from \code{MLEbin.simulate()}; see
\code{?MLEbin.simulate()} for details}

\item{vertCol}{colour for vertical line at true value of \code{b}}

\item{vertThick}{thickness of vertical line at true value of \code{b}}

\item{xrange}{range of x values TODO can change to xlimA for consistency}

\item{xbigticks.by}{increment between big tick marks on x-axis (all labelled)}

\item{xsmallticks.by}{increment between small tick marks on x-axis}

\item{omi, mfrow, mai, xaxs, yaxs, mgp, cex, inset}{standard options for \code{par()},
defaults are for Figure 4}

\item{insetMat}{matrix of inset values for MLEmid column (Figure 5(e) legend
has to be shifted), each row is the inset for that MLEmid panel}

\item{legLabMid}{character vector of length 4 for labelling each panel in
the MLEmid column}

\item{legLabBin}{character vector of length 4 for labelling each panel in
the MLEBin column}
}
\value{
figure with eight panels of confidence intervals, one for each combination of binning
type and fitting method.
}
\description{
The default plot here reproduces Figure 5 of MEPS paper, showing the
confidence intervals of the exponent $b$ for 10,000 simulated data sets, binned using four
methods and fitted using the MLEmid and MLEbin methods.
}
\author{
Andrew Edwards
}
