% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{MLEmid.MLEbin.hist}
\alias{MLEmid.MLEbin.hist}
\title{One figure with eight histograms for each of  MLEmid and MLEbin methods and four binning types}
\usage{
MLEmid.MLEbin.hist(
  results.list,
  vertCol = "red",
  vertThick = 1,
  xrange = c(-2.4, -1.1),
  xbigticks.by = 0.4,
  xsmallticks.by = 0.1,
  ylimA = c(0, 7000),
  yBigTickLab.by = 3000,
  yBigTickNoLab.by = 1000,
  ySmallTick.by = 500,
  binwidth = NA,
  omi = c(0.12, 0.05, 0.2, 0),
  mfrow = c(4, 2),
  mai = c(0.5, 0.5, 0, 0.3),
  xaxs = "i",
  yaxs = "i",
  mgp = c(2, 0.5, 0),
  cex = 0.8,
  cexAxis = 0.9,
  inset = c(0, -0.04),
  legLabMid = c("(a)", "(c)", "(e)", "(g)"),
  legLabBin = c("(b)", "(d)", "(f)", "(h)")
)
}
\arguments{
\item{results.list}{output list from \code{MLEbin.simulate()}; see
\code{?MLEbin.simulate()} for details}

\item{vertCol}{colour for vertical line at true value of \code{b}}

\item{vertThick}{thickness of vertical line at true value of \code{b}}

\item{xrange}{range of x values (should change to xlimA for consistency)}

\item{xbigticks.by}{increment between big tick marks on x-axis (all labelled)}

\item{xsmallticks.by}{increment between small tick marks on x-axis}

\item{ylimA}{range of y values}

\item{yBigTickLab.by}{increment between labelled big tick marks on y-axis}

\item{yBigTickNoLab.by}{increment between all big tick marks on y-axis}

\item{ySmallTick.by}{increment between small tick marks on y-axis}

\item{binwidth}{bin width for histograms; if NA then gets calculated such
that \code{b.known} is a midpoint of a bin}

\item{omi, mfrow, mai, xaxs, yaxs, mgp, cex, inset}{standard options for \code{par()},
defaults are for Figure 4}

\item{cexAxis}{font size for axis labels}

\item{legLabMid}{character vector of length 4 for labelling each panel in
the MLEmid column}

\item{legLabBin}{character vector of length 4 for labelling each panel in
the MLEBin column}
}
\value{
figure with eight histograms, one for each combination of binning
type and fitting method.
}
\description{
The default plot here reproduces Figure 4 of MEPS paper, showing the
estimated exponent $b$ for 10,000 simulated data sets, binned using four
methods and fitted using the MLEmid and MLEbin methods.
}
\author{
Andrew Edwards
}
