% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{MLEmid.MLEbin.table}
\alias{MLEmid.MLEbin.table}
\title{Make dataframe of results from MLEmid and MLEbin methods and four binning types}
\usage{
MLEmid.MLEbin.table(results.list)
}
\arguments{
\item{results.list}{output list from \code{MLEbin.simulate()}; see
\code{?MLEbin.simulate()} for details}
}
\value{
dataframe of table with columns \code{Binning.type}, \code{Method},
\code{Quantile.5}, \code{Median, }Mean\code{, }Quantile.95\code{, }Percent.below.true`, one
row for each combination of binning type and fitting method.
}
\description{
Makes dataframe to produce Tables S.3, S.4 and S.5 of MEPS paper, showing the
statistics from estimating exponent $b$ for 10,000 simulated data sets, binned using four
methods and fitted using the MLEmid and MLEbin methods.
}
\author{
Andrew Edwards
}
