% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{mode_lm_res_beta_Secondmode}
\alias{mode_lm_res_beta_Secondmode}
\title{modal analysis linear regression}
\usage{
mode_lm_res_beta_Secondmode(dat)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
modal analysis linear regression
}
