% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{mode_vs_mode_lm}
\alias{mode_vs_mode_lm}
\title{modal analysis scatterplot plus regression line}
\usage{
mode_vs_mode_lm(dat)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
modal analysis scatterplot plus regression line
}
