% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_response_variables.R
\name{modes_pelagic}
\alias{modes_pelagic}
\title{Make pelagic modes}
\usage{
modes_pelagic(dat, dat2)
}
\arguments{
\item{dat}{}
}
\value{
dat3
}
\description{
Make pelagic modes
}
\examples{



}
