% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{multi_covariate_marg_bruvs}
\alias{multi_covariate_marg_bruvs}
\title{multi marginal plot across models}
\usage{
multi_covariate_marg_bruvs(
  beta_marg,
  firstmode_marg,
  secondmode_marg,
  covariate_name
)
}
\arguments{
\item{beta_marg}{marg betaslop}

\item{firstmode_marg}{marg firstmode}

\item{secondmode_marg}{marg secondmode}

\item{covariate_name}{name of covariate}
}
\value{

}
\description{
multi marginal plot across models
}
