% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{multi_covariate_marg_SST_sd}
\alias{multi_covariate_marg_SST_sd}
\title{multi marginal plot across models only SST_sd}
\usage{
multi_covariate_marg_SST_sd(beta_marg, firstmode_marg, covariate_name)
}
\arguments{
\item{beta_marg}{marg betaslop}

\item{firstmode_marg}{marg firstmode}

\item{covariate_name}{name of covariate}
}
\value{

}
\description{
multi marginal plot across models only SST_sd
}
