% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{multi_envar_range}
\alias{multi_envar_range}
\title{plot predictor variables for each bruvs type}
\usage{
multi_envar_range()
}
\arguments{
\item{var_logBathy}{bahty}

\item{var_logDistCR}{distance to coral reef}

\item{var_logDistSM}{seamount}

\item{var_logDistP}{port}

\item{var_logDistC}{coast}

\item{var_logTTM}{travel time marke}

\item{var_logCHL}{chal}

\item{var_logSST}{sst_mean}

\item{var_SST_sd}{sst_SD}

\item{var_SST_GEm}{goverment effectivness}
}
\value{

}
\description{
plot predictor variables for each bruvs type
}
