% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{multi_envar_range_cat}
\alias{multi_envar_range_cat}
\title{plot predictor variables for each bruvs type plus protection}
\usage{
multi_envar_range_cat()
}
\arguments{
\item{prot_var_logBathy}{bahty}

\item{prot_var_logDistCR}{distance to coral reef}

\item{prot_var_logDistSM}{seamount}

\item{prot_var_logDistP}{port}

\item{prot_var_logDistC}{coast}

\item{prot_var_logTTM}{travel time marke}

\item{prot_var_logCHL}{chal}

\item{prot_var_logSST}{sst_mean}

\item{prot_var_SST_sd}{sst_SD}

\item{prot_var_SST_GEm}{goverment effectivness}
}
\value{

}
\description{
plot predictor variables for each bruvs type plus protection
}
