% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{multi_fig_sample}
\alias{multi_fig_sample}
\title{multiplot fig 1 sampling effort and rank order species}
\usage{
multi_fig_sample(fig_map, fig_sp_rank)
}
\arguments{
\item{covariate_name}{name of covariate}
}
\value{

}
\description{
multiplot fig 1 sampling effort and rank order species
}
