% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{negLL.PLB.binned.species}
\alias{negLL.PLB.binned.species}
\title{Calculate the negative log-likelihood of \code{b} for the PLB model, given
species-specific binned data (MLEbins method). DEPRECATED -- \code{use negLL.PLB.bins.species()}}
\usage{
negLL.PLB.binned.species(b, dataBinForLike, dataBinForLikeSummary)
}
\arguments{
\item{b}{value of \code{b} for which to calculate the negative log-likelihood}

\item{dataBinForLike}{table data frame (tbl_df) where each row is the count in a bin
of a species, and columns (and corresponding mathematical notation in MEPS Appendix) are:
\itemize{
\item \code{SpecCode}: code for each species, \code{s}
\item \code{wmin}: lower bound of the bin, \code{w_\{sj\}} where \code{j} is the bin number
\item \code{wmax}: upper bound of the bin, \code{w_\{s,j+1\}}
\item \code{Number}: count in that bin for that species, \code{d_\{sj\}}
For each species the first and last bins must be non-empty, i.e.
\code{w_\{s1\}, w_\{s,J_s +1\} > 0}. Should ideally write code to check that before
calling this function (since this gets repeatedly called).
}}

\item{dataBinForLikeSummary}{tbl_df with one row for each species, giving
the minimum lower bound \code{w_\{s1\}} and maximum upper bound \code{w_\{s,J_s +1\}}
and the number of counts for that species, where \code{J_s} is the number of
bins for species \code{s} (won't need to explicilty specify). Columns are:
\itemize{
\item \code{SpecCode}: code for each species \code{s}
\item \code{wminSpecies}: minimum lower bound \code{w_\{s1\}}
\item \code{wmaxSpecies: maximum upper bound }w_{s,J_s +1}`
\item \code{n_s}: total number of counts for species \code{s} \code{n_s}
}}
}
\value{
negative log-likelihood of the parameters given the data
}
\description{
Calculate the negative log-likelihood of \emph{b} for the PLB model,
given binned data where the bins can be different for each species.
Returns the negative log-likelihood. This was Andy's original likelihood
function but we replaced it with Mike's simpler one, so use
\code{negLL.PLB.bins.species()}. Documented this one first, so keeping.
Will be called by \code{nlm()} or similar, but \code{xmin} and \code{xmax} will just be estimated
as the min of lowest bin and max of the largest bin (i.e. their MLEs),
no need to do numerically.
}
\author{
Andrew Edwards
}
