% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{negLL.PLB.bins.species}
\alias{negLL.PLB.bins.species}
\title{TODO DOCUMENT ME IF NECESSARY - YES, is called from MLEbins vignette}
\usage{
negLL.PLB.bins.species(b, dataBinForLike, n, xmin, xmax)
}
\arguments{
\item{b}{temp}

\item{dataBinForLike}{temp}

\item{n}{temp}

\item{xmin}{temp}

\item{xmax}{temp}
}
\value{
temp
}
\description{
TODO Need to clarify how this differs from \code{negLL.PLB.binned.species()}
}
\author{
Andrew Edwards
}
