% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{negLL.PLB}
\alias{negLL.PLB}
\title{Calculate negative log-likelihood for the bounded power-law
distribution}
\usage{
negLL.PLB(b, x, n, xmin, xmax, sumlogx)
}
\arguments{
\item{b}{value of \code{b} for which to calculate the negative log-likelihood}

\item{x}{vector of values of data (e.g. masses of individual fish)}

\item{n}{\code{length(x)}, have as an input to avoid repeatedly calculating it when
function is called multiple times in an optimization routine}

\item{xmin}{minimum value of \code{x} to avoid repeatedly calculating}

\item{xmax}{maximum value of \code{x} to avoid repeatedly calculating}

\item{sumlogx}{\code{sum(log(x))} to avoid repeatedly calculating}
}
\value{
negative log-likelihood of the parameters given the data
}
\description{
Calculate the negative log-likelihood of the parameters \code{b}, \code{xmin} and
\code{xmax} given data \code{x} for the PLB model. Returns the negative
log-likelihood. Will be called by \code{nlm()} or similar. \code{xmin} and \code{xmax}
are just estimated as the min and max of the data, not numerically using likelihood.
}
\author{
Andrew Edwards
}
