% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pBiomassBins}
\alias{pBiomassBins}
\title{Total and normalised biomass in each bin for a fitted distribution and given
bin breaks}
\usage{
pBiomassBins(
  ...,
  binValsTibble = NULL,
  binBreaks = NULL,
  xmin = NULL,
  xmax = NULL,
  n
)
}
\arguments{
\item{...}{extra arguments passed to \code{bBiomass}: \code{b}.
Needs to go first to avoid
partial matching of \code{b} (to be part of \code{...}) with whichever of
\code{binValsTibble} or \code{binBreaks} is not supplied in the call to
\code{pBiomassBins()}. Will likely want to specify these in every call (do not
rely on the default).}

\item{binValsTibble}{tibble of binned data with each row representing a bin
and with columns \code{wmin} and \code{wmax} (min and max break of each bin), or
columns named \code{binMin} and \code{binMax}.
Extra columns are ignored. Similar to \code{LBN_bin_plot()}.}

\item{binBreaks}{vector of bin breaks}

\item{n}{total number of individuals in the system (needed to calculate biomass)}
}
\value{
tibble with each row corresponding to a bin, and columns \code{wmin},
\code{wmax},  \code{binWidth}, \code{estBiomass}, and \code{estBiomassNorm}. If the input is a tibble with
columns \code{wmin} and \code{wmax} (or \code{binMin} and \code{binMax}), then
columns \code{estBiomass} and \code{estBiomassNorm} are appended to the input tibble.
}
\description{
Bin breaks are input as EITHER a single tibble \code{binValsTibble}
with each row representing a bin, OR as a vector \code{binBreaks} of breaks.
}
\author{
Andrew Edwards
}
