% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_binned_fitted}
\alias{plot_binned_fitted}
\title{Add horizontal bars and shaded rectangles to \code{LBN_bin_plot()}}
\usage{
plot_binned_fitted(
  binTibble,
  bar.col = "red",
  bar.lwd = 3,
  rect.shading = "pink",
  shorter = 0.05
)
}
\arguments{
\item{binTibble}{tibble of values calculated in \code{LBN_bin_plot()}; this
function adds to that plot}

\item{bar.col}{colour for the horizontal bars representing the MLE values of
normalised biomass in each bin}

\item{bar.lwd}{thickness of horiztonal bars}

\item{rect.shading}{colour for shading of rectangles corresponding to the
normalised biomasses estimated from confidence intervals of \code{b}}

\item{shorter}{fraction shorter to make the rectangles, so can see them
overlapping with grey rectangles; may not work exacly as planned (won't be symmetric) when x-axis
not logged, but that's not going to be a useful plot anyway}
}
\value{

}
\description{
These are to show the estimated normalised biomasses in each bin, based on
the MLE value of \code{b} and it's confidence limit values.
Each horizontal bar spans a bin (default colour is red), with it's vertical value indicating the
expected normalised biomass based on the MLE of \code{b}. The height of the
shaded rectangles (default colour pink) show the range of expected
normalised biomass based on the 95\\% confidence interval of \code{b}.
}
\author{
Andrew Edwards
}
