% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{profLike}
\alias{profLike}
\title{Profile log-likelihood method to calculate 95\% confidence interval}
\usage{
profLike(negLL.fn, MLE, minNegLL, vecDiff = 0.5, vecInc = 0.001, ...)
}
\arguments{
\item{negLL.fn}{negative log-likelihood function that take arguments
(parameters and data) in ... and returns a negative log-likelihood value}

\item{MLE}{maximum likelihood estimate (already calculated)}

\item{minNegLL}{the minimum of the negative log-likelihood function, at the
MLE (by definition)}

\item{vecDiff}{value defining  range over which to test the negative log-likelihood
to construct the confidence interval; range is \code{MLE} \eqn{\pm} \code{vecDiff}. Default is 0.5 and a symmetric
range is tested for fitting size spectra, since for movement data
sets in Table 2 of Edwards (2011; 92(6):1247-1257) the intervals were
symmetric, so symmetric seems a good start.}

\item{vecInc}{increments to try, the accuracy of the resulting bounds
will depend on this. Note that a resulting interval of, say,
(-2.123, -1.987) means that that interval is contained within the
true 95\% interval, which is itself contained within (-2.124, -1.986).
The true bounds lie between the stated lower bounds and between
the stated upper bounds. So reduce vecInc if further accuracy is needed.}

\item{...}{further arguments to \code{negLL.fn()}}
}
\value{
two-component vector of the 95\% confidence interval
}
\description{
Profile log-likelihood method to calculate 95\% confidence interval
for a given negative log-likelihood function, maximum likelihood estimate
(MLE) and minimum of of the negative log-likelihood function. Based on
Hilborn and Mangel (1997), The Ecological Detective, p162.
}
\author{
Andrew Edwards
}
