% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{rbind_fl_meta}
\alias{rbind_fl_meta}
\title{rbind benthic and pelagic lengths with meta}
\usage{
rbind_fl_meta(dat_pelagic_fl_meta, dat_benthic_fl_meta)
}
\arguments{
\item{dat_pelagic_fl_meta}{...}

\item{dat_benthic_fl_meta}{...}
}
\description{
rbind benthic and pelagic lengths with meta
}
