% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{rbind_meta_coordinates}
\alias{rbind_meta_coordinates}
\title{rbind pelagic and benthic meta coordinates}
\usage{
rbind_meta_coordinates(dat_pelagic, dat_benthic)
}
\arguments{
\item{dat_pelagic}{...}

\item{dat_benthic}{...}
}
\description{
rbind pelagic and benthic meta coordinates
}
