% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{read_data_pelagic_benthic_sum}
\alias{read_data_pelagic_benthic_sum}
\title{read pelagic benthic summary  location}
\usage{
read_data_pelagic_benthic_sum()
}
\value{

}
\description{
read pelagic benthic summary  location
}
