% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{read_marg_pred_all_sens}
\alias{read_marg_pred_all_sens}
\title{read marginal sensitivity predictions}
\usage{
read_marg_pred_all_sens()
}
\value{

}
\description{
read marginal sensitivity predictions
}
