% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{read_marg_pred_all_sens_oceans}
\alias{read_marg_pred_all_sens_oceans}
\title{read marginal sensitivity predictions oceans}
\usage{
read_marg_pred_all_sens_oceans()
}
\value{

}
\description{
read marginal sensitivity predictions oceans
}
