% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{remove_single_opcodes}
\alias{remove_single_opcodes}
\title{remove from fl the opcodes not available in maxn, ie the single opcodes, for pelagic or benthic bruvs}
\usage{
remove_single_opcodes(dat_fl, dat_maxn)
}
\arguments{
\item{dat_fl}{}

\item{dat_maxn}{}
}
\value{

}
\description{
remove from fl the opcodes not available in maxn, ie the single opcodes, for pelagic or benthic bruvs
}
