% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{remove_single_opcodes_pelagic}
\alias{remove_single_opcodes_pelagic}
\title{remove from a data (fl or maxn) the opcode not available in the other data (maxn or fl), ie the single opcodes, for pelagic bruvs}
\usage{
remove_single_opcodes_pelagic(dat1, dat2)
}
\arguments{
\item{dat1}{}

\item{dat2}{}
}
\value{

}
\description{
remove from a data (fl or maxn) the opcode not available in the other data (maxn or fl), ie the single opcodes, for pelagic bruvs
}
