% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{removed_notpermitted_expeditions_benthic}
\alias{removed_notpermitted_expeditions_benthic}
\title{Removed expeditions that are not permitted to be used in this study for benthic bruvs}
\usage{
removed_notpermitted_expeditions_benthic(dat, expeds_notpermit)
}
\arguments{
\item{dat}{}

\item{expeds_permit}{}
}
\value{

}
\description{
Removed expeditions that are not permitted to be used in this study for benthic bruvs
}
