% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{response_fig}
\alias{response_fig}
\title{response variable multiplot}
\usage{
response_fig(
  dat2,
  dat,
  tab_first,
  tab_second,
  min_size,
  lat_band,
  bandw,
  scale,
  alpha
)
}
\arguments{
\item{dat}{}
}
\value{

}
\description{
response variable multiplot
}
