% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{response_fig_new}
\alias{response_fig_new}
\title{response variable multiplot}
\usage{
response_fig_new(dat, tab_first, tab_second, min_size)
}
\arguments{
\item{dat}{}
}
\value{

}
\description{
response variable multiplot
}
