% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sizespectra.R
\name{response_globalmap}
\alias{response_globalmap}
\title{Global map of response variables}
\usage{
response_globalmap(world, mar, meta_pb)
}
\arguments{
\item{world}{the output of \code{ggplot2::map_data()}}

\item{mar}{an sp object: world marine boundaries}

\item{meta_pb}{coordinates of...}
}
\value{
A ggplot2 object
}
\description{
Global map of response variables
}
\examples{
## ...
}
