% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_dplyr.R
\name{s_select}
\alias{s_select}
\title{Modified version of dplyr's select that uses string arguments}
\usage{
s_select(.data, ...)
}
\arguments{
\item{.data}{input}

\item{...}{other arguments to \code{dplyr::select()}}
}
\description{
Taken from https://gist.github.com/skranz/9681509 by Sebastian Kranz
}
\author{
Sebastian Kranz (https://gist.github.com/skranz/9681509)
}
