% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{save_marg_pred_all_sens}
\alias{save_marg_pred_all_sens}
\title{save marginal sensitivity predictions}
\usage{
save_marg_pred_all_sens(pred_all)
}
\value{

}
\description{
save marginal sensitivity predictions
}
