% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{save_marg_pred_all_sens_oceans}
\alias{save_marg_pred_all_sens_oceans}
\title{save marginal sensitivity predictions oceans}
\usage{
save_marg_pred_all_sens_oceans(pred_all)
}
\value{

}
\description{
save marginal sensitivity predictions oceans
}
