% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{select_fish_families_benthic}
\alias{select_fish_families_benthic}
\title{select fish families in benthic bruvs (fl/maxn data)}
\usage{
select_fish_families_benthic(dat, families)
}
\arguments{
\item{dat}{}

\item{families}{}
}
\value{

}
\description{
select fish families in benthic bruvs (fl/maxn data)
}
