% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{select_fish_families_pelagic}
\alias{select_fish_families_pelagic}
\title{select fish families in pelagic bruvs (fl/maxn data)}
\usage{
select_fish_families_pelagic(dat, families)
}
\arguments{
\item{dat}{}

\item{families}{}
}
\description{
select fish families in pelagic bruvs (fl/maxn data)
}
