\name{SizeSpectra}
\alias{SizeSpectra}
\docType{package}
\title{
Marine Size-Spectra Package
}
\description{
Spatial Size-spectra analysis functions for aquatic ecosystems.
}
\details{
\tabular{ll}{
Package: \tab SizeSpectra\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-07-11\cr
License: \tab What license is it under?\cr
}
This package contains various functinos for creating plotting and analysing 
spatially explicit predation driven marine size spectra. It can calculate 
and analyse size spectra in 0, 1 and 2 spatial dimensions, and for any number of 
distinct species.
\item\code{Setup.Grid}
\item\code{Setup.Run}
\item\code{Setup.Species}
\item\code{Setup.Initial}
\item\code{SizeSpectrum}
\item\code{Read.In}
\item\code{Extract.Time}
\item\code{Average.Time}
\item\code{Plot.Spectrum}
\item\code{Animate}
}
\author{
Matt castle

Maintainer:  <mattcastle@cefas.co.uk>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
