% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{specCodeNames}
\alias{specCodeNames}
\title{Species codes and their scientific names}
\format{Tibble with columns
\itemize{
\item species: Scientific name of species
\item speccode: Species code used in the IBTS data for that species (maybe
not all species codes are here).
}}
\source{
Aphia species codes, Julia obtained them from DATRAS (the ICES
DATRAS R packages likely has them all).
}
\usage{
specCodeNames
}
\description{
Species codes and their scientific names
}
\keyword{datasets}
