% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{species_bins_plots}
\alias{species_bins_plots}
\title{Plot the species-specific bins for all species, as in Figures 6, S.1, S.2 and S.3}
\usage{
species_bins_plots(
  dataBin_vals = dataBin,
  postscript = FALSE,
  specCodeHighlight = c(127205, 154675)
)
}
\arguments{
\item{dataBin_vals}{dataframe in the same format as the saved dataset \code{dataBin}
-- see \code{?dataBin} for structure}

\item{postscript}{If TRUE then save four separate postscript figures, else
just plot the four figures to the active device}

\item{specCodeHighlight}{Species to highlight in red in Figure 6 (species
127251 for Figure A.4 is done within the function).}
}
\value{
list containing:
\itemize{
\item maxWidth: maximum width of any bin
\item specNameMaxWidth: species name corresponding to \code{maxWidth}
\item maxRatio: maximum ratio of a body-mass bin to it's minimum, i.e. maximum
value of \code{(wmax - wmin)/wmin}
\item specNameMaxRatio: species name corresponding to \code{specNameMaxRatio}
}
}
\description{
Wrangles the data and then plots the Figures 6, S.1, S.2 and S.3 showing how
the species-specific length-weight coefficients yield different bins.
}
\author{
Andrew Edwards
}
