% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{stepAIC_mod}
\alias{stepAIC_mod}
\title{Title}
\usage{
stepAIC_mod(
  object,
  scope,
  scale = 0,
  direction = c("both", "backward", "forward"),
  trace = 1,
  keep = NULL,
  steps = 1000,
  use.start = FALSE,
  k = 2,
  ...
)
}
\arguments{
\item{object}{}

\item{scope}{}

\item{scale}{}

\item{direction}{}

\item{trace}{}

\item{keep}{}

\item{steps}{}

\item{use.start}{}

\item{k}{}

\item{...}{}
}
\value{

}
\description{
Title
}
