% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{timeSerPlot.eight}
\alias{timeSerPlot.eight}
\title{Call \code{timeSerPlot()} eight times to create MEPS Figure 1, eight methods
applied to 30 years of data}
\usage{
timeSerPlot.eight(fullResults.local = fullResults)
}
\arguments{
\item{fullResults.local}{Results from applying eight methods to a
dataset. Must be in same format as \code{fullResults} which is for the IBTS
data. Not tested yet with other data sets.}
}
\value{
Dataframe with one row for each of the main eight methods, where
each row is an output from \code{timeSerPlot()} and so has columns:
\itemize{
\item Method: method used
\item Low: lower bound of 95\\% confidence interval
\item Trend: gradient of regression fit
\item High: upper bound of 95\\% confidence interval
\item p: p-value of regression fit
\item Rsquared: r-squared of regression fit
\item adjRsquared: adjusted r-squared of regression fit
}
}
\description{
For each of the eight methods in turn, call \code{timeSerPlot()} to add a panel
with the estimated $b$ and 95\\% confidence interval for the method for each
year of IBTS data, and then fit a regression through the estimates. Called
in vignette \code{MEPS_IBTS_2.Rmd}.
}
\author{
Andrew Edwards
}
