% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{timeSerPlot}
\alias{timeSerPlot}
\title{Plot time series of estimated \emph{b} with confidence intervals}
\usage{
timeSerPlot(
  bForYears,
  legName,
  method,
  weightReg = FALSE,
  bCol = "black",
  pchVal = 20,
  cexVal = 1,
  confCol = "black",
  confThick = 1,
  xLim = NULL,
  yLim = NULL,
  xLab = "",
  yLab = expression(paste("Estimate of ", italic(b)), sep = ""),
  xTicksSmallInc = NULL,
  xTicksSmallTck = 0.01,
  yLabels = TRUE,
  yTicksSmallInc = NULL,
  yTicksSmallTck = 0.01,
  legPos = "topleft",
  newPlot = TRUE,
  regPlot = TRUE,
  regColNotSig = "darkgrey",
  regColSig = "red",
  legExtra = NULL,
  legExtraPos = "topleft",
  legExtraCol = "",
  insetVal = c(-0.08, -0.06),
  xJitter = 0,
  doRegression = TRUE
)
}
\arguments{
\item{bForYears}{dataframe with columns \code{Year}, \code{Method} (optional), \code{b}
(the estimate of \emph{b}), \code{confMin} and \code{confMax} (the 95\\% lower and upper
confidence limits) and \code{stdError} (the standard error of the estimate of
\emph{b}).}

\item{legName}{legend name for that panel}

\item{method}{method used to obtain the inputted estimates of \code{b}}

\item{weightReg}{TRUE if doing weighted regression (using standard errors)
or FALSE to not do weighted regression.}

\item{bCol}{colour for points for \emph{b}}

\item{pchVal}{pch for points for \emph{b}}

\item{cexVal}{size of points for \emph{b}}

\item{confCol}{colour for confidence intervals for \emph{b}}

\item{confThick}{thickness of vertical line for confidence intervals}

\item{xLim}{x-axis limits}

\item{yLim}{y-axis limits}

\item{xLab}{label for x-axis}

\item{yLab}{label for y-axis}

\item{xTicksSmallInc}{increments for where to have small (unlabelled)
tickmarks on x-axis}

\item{xTicksSmallTck}{tick length for small (unlabelled) tickmarks on x-axis}

\item{yLabels}{whether or not to label main tickmarks on y-axis}

\item{yTicksSmallInc}{increments for where to have small (unlabelled)
tickmarks on y-axis}

\item{yTicksSmallTck}{tick length for small (unlabelled) tickmarks on y-axis}

\item{legPos}{legend position}

\item{newPlot}{TRUE to create a new plot, FALSE to add to existing}

\item{regPlot}{TRUE to plot the regression line and conf intervals}

\item{regColNotSig}{colour for regression line (and its confidence intervals)
if the trend is not significant}

\item{regColSig}{colour for regression line (and its confidence intervals)
if the trend is significant}

\item{legExtra}{extra manually-specified legend (e.g. to distinguish two
sets of results)}

\item{legExtraPos}{position for extra manually-specified legend}

\item{legExtraCol}{colours (vector) for extra manually-specified legend}

\item{insetVal}{inset shift for naming the panel}

\item{xJitter}{value to jitter the x-values by (for comparison plot the
confidence intervals overlap)}

\item{doRegression}{whether to calculate a regression for the time series of estimates}
}
\value{
if \code{doRegression} is TRUE (default) then return dataframe of just
one row with columns (else return nothing):
\itemize{
\item Method: method used
\item Low: lower bound of 95\\% confidence interval
\item Trend: gradient of regression fit
\item High: upper bound of 95\\% confidence interval
\item p: p-value of regression fit
\item Rsquared: r-squared of regression fit
\item adjRsquared: adjusted r-squared of regression fit
}
}
\description{
Plot time series of estimated \emph{b} with confidence intervals, for
data that are analysed year-by-year by a single method. And then
fit a linear regression with its confidence intervals (or not if
\code{doRegression = FALSE}).
This will get called eight times to produce a comparison figure of
the methods.
}
\author{
Andrew Edwards
}
