% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trendResults}
\alias{trendResults}
\title{Regression fits from using each fitting method on each year of IBTS data set
(for MEPS Table S.1)   TODO maybe not needed since calculated in vignette}
\format{A dataframe with one row for each of the main eight methods, where
each row is an output from \code{timeSerPlot()} and so has columns:
\itemize{
\item Method: method used
\item Low: lower bound of 95\% confidence interval
\item Trend: gradient of regression fit
\item High: upper bound of 95\% confidence interval
\item p: p-value of regression fit
\item Rsquared: r-squared of regression fit
\item adjRsquared: adjusted r-squared of regression fit
}}
\source{
Generated from running vignette \code{MEPS_IBTS_2}.
}
\usage{
trendResults
}
\description{
Regression fits from using each fitting method on each year of IBTS data set
(for MEPS Table S.1)   TODO maybe not needed since calculated in vignette
}
\keyword{datasets}
