% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{truncate_data}
\alias{truncate_data}
\title{truncate predictor variables to three standard deviation from the mean}
\usage{
truncate_data(dat, response_name, std_no)
}
\arguments{
\item{dat}{dataframe with predictor variables}

\item{response_name}{response name}
}
\value{

}
\description{
truncate predictor variables to three standard deviation from the mean
}
