% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{vif_mod}
\alias{vif_mod}
\title{Estimate variance inflation factor and visualise the VIF values}
\usage{
vif_mod(mod)
}
\arguments{
\item{mod}{saturated model to test for VIF}
}
\value{

}
\description{
Estimate variance inflation factor and visualise the VIF values
}
