% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{write_merged_fl_meta}
\alias{write_merged_fl_meta}
\title{save merged meta with FL (benthic or pelagic)}
\usage{
write_merged_fl_meta(dat, type)
}
\arguments{
\item{dat}{...}

\item{type}{...}
}
\value{

}
\description{
save merged meta with FL (benthic or pelagic)
}
