% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{write_meta_opcodes_benthic}
\alias{write_meta_opcodes_benthic}
\title{Write meta opcodes for benthic bruvs}
\usage{
write_meta_opcodes_benthic(dat)
}
\arguments{
\item{dat}{}
}
\value{

}
\description{
Write meta opcodes for benthic bruvs
}
