% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{write_meta_opcodes_pelagic}
\alias{write_meta_opcodes_pelagic}
\title{Write meta opcodes for pelagic bruvs}
\usage{
write_meta_opcodes_pelagic(dat)
}
\arguments{
\item{dat}{}
}
\value{

}
\description{
Write meta opcodes for pelagic bruvs
}
