% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{write_meta_strings_pelagic}
\alias{write_meta_strings_pelagic}
\title{Write meta strings for pelagic bruvs}
\usage{
write_meta_strings_pelagic(dat)
}
\arguments{
\item{dat}{}
}
\description{
Write meta strings for pelagic bruvs
}
