% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{clean_meta_benthic}
\alias{clean_meta_benthic}
\title{Clean meta data for benthic bruvs}
\usage{
clean_meta_benthic(dat)
}
\arguments{
\item{dat}{}
}
\description{
Clean meta data for benthic bruvs
}
