% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{eightMethodsMEE}
\alias{eightMethodsMEE}
\title{Use all eight methods to fit a simple vector of body masses as in the MEE paper}
\usage{
eightMethodsMEE(x, num.bins = 8, b.only = FALSE)
}
\arguments{
\item{x}{vector of individual body masses}

\item{num.bins}{suggested number of bins for Llin, LT and LTplus1 methods}

\item{b.only}{TRUE only returns slope or b plus confidence intervals, else
return full details}
}
\value{
If \code{b.only} is TRUE then return a list with each element (one for
each method) being a
vector the slope or b followed by its confidence interval.
If \code{b.only} is FALSE then return list of full results (need to specify
here -- see example in vignette \code{MEE_reproduce_1}).
}
\description{
Use all eight of the methods described in MEE paper to fit size spectra to a
vector of body masses. Data could be lengths but then the LBmiz, LBbiom and
LBNbiom methods are meaningless.
}
\author{
Andrew Edwards
}
