% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{histAxes}
\alias{histAxes}
\title{Histogram axes for figures TODO}
\usage{
histAxes(yBigTickLab = c(0, 2000, 4000), yBigTickNoLab = seq(0, 6500,
  1000), ySmallTick = seq(0, 6500, 500))
}
\arguments{
\item{yBigTickLab}{y-axis big ticks to label}

\item{yBigTickNoLab}{y-axis big ticks to not label}

\item{ySmallTick}{y-axis small ticks (unlabelled)}
}
\value{
Adds axes to existing histogram TODO: check
}
\description{
Do the histogram axes in \code{fitting3rep.r} TODO and later, since almost all
panels will have same axes. Not very flexible, use \code{histAxes2()}
to plot up to 10,000. TODO Note that \code{xbsmallticks} and \code{xbigticks} need to be
prespecified, but
they haven't yet been made arguments here.
}
\author{
Andrew Edwards
}
