% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qqtab}
\alias{qqtab}
\title{Produce LaTeX table code for quantiles of results}
\usage{
qqtab(xx, dig = 2, true = b.known, quants = c(0.25, 0.75))
}
\arguments{
\item{xx}{vector of values to give quantiles for}

\item{dig}{number of decimal places to give}

\item{true}{the true value of the quantity being estimated, will
depend on method}

\item{quants}{quantiles to use, with defaults of 0.25 and 0.75 (25\% and 75\%).}
}
\value{
LaTeX output for table to copy into a .tex file
}
\description{
Quantile table function, to construct lines of quantiles to copy
into LaTeX. Does \code{quants[1]}, 50\% (median), mean, \code{quants[2]} quantiles,
and the \%age of values < the true value.
}
\author{
Andrew Edwards
}
